+++ .github/workflows/update.yml
@@ -2,16 +2,28 @@ name: Up to Date
 
 on:
   schedule:
-    - cron: '0 0 * * *'
+    - cron: "0 0 * * *"
 
 jobs:
   update:
     name: Update
     runs-on: ubuntu-latest
     steps:
-      - uses: actions/checkout@v3
-      - uses: ./.github/actions/setup
-      - run: pnpm run generate
+      - name: Checkout Repository
+        uses: actions/checkout@v2
+        with:
+          fetch-depth: 0
+
+      - name: Setup Node.js
+        uses: actions/setup-node@v2-beta
+        with:
+          node-version: "12"
+
+      - name: Install Dependencies
+        run: yarn
+
+      - name: Generate Cheat Sheet
+        run: yarn generate
 
       - name: Push New Cheat Sheet to Updated Branch
         run: |
@@ -20,7 +32,7 @@ jobs:
           git config --global user.name "github-actions[bot]"
           git config --global user.email "41898282+github-actions[bot]@users.noreply.github.com"
           git commit -m "docs(readme): update emoji-cheat-sheet"
-          git push -qf "https://${{ secrets.GITHUB_TOKEN }}@github.com/${{ github.repository }}.git" HEAD:github-actions-auto-update
+          git push -qf "https://${{ secrets.GITHUB_TOKEN }}@github.com/ikatyang/emoji-cheat-sheet.git" HEAD:github-actions-auto-update
 
       - name: Create PR from Updated Branch
         uses: actions/github-script@v2
