+++ .gitignore
@@ -1,2 +1,1 @@
-/.vscode/
-/node_modules/
+node_modules/
