--- .prettierignore
@@ -1,2 +1,0 @@
-/pnpm-lock.yaml
-/README.md
