--- .prettierrc.yml
@@ -1,3 +1,0 @@
-arrowParens: avoid
-semi: false
-singleQuote: true
