+++ package.json
@@ -1,6 +1,5 @@
 {
   "name": "emoji-cheat-sheet",
-  "type": "module",
   "version": "0.0.0-dev",
   "private": true,
   "repository": "ikatyang/emoji-cheat-sheet",
@@ -12,21 +11,27 @@
   },
   "license": "MIT",
   "scripts": {
-    "lint": "prettier --check .",
-    "check": "tsc --noEmit",
-    "test": "vitest",
-    "generate": "vite-node ./scripts/generate.ts -- run > ./README.md"
+    "lint": "tslint -p . --type-check",
+    "test": "jest",
+    "generate": "node ./scripts/generate.js > ./README.md"
+  },
+  "dependencies": {
+    "cheerio": "^0.22.0",
+    "request": "^2.88.0"
   },
   "devDependencies": {
-    "jest-playback": "4.0.0",
-    "prettier": "3.0.0",
-    "typescript": "5.1.6",
-    "vite": "4.5.2",
-    "vite-node": "0.33.0",
-    "vitest": "0.33.0"
+    "@types/cheerio": "0.22.13",
+    "@types/jest": "24.0.18",
+    "@types/node": "12.7.4",
+    "@types/request": "2.48.2",
+    "jest": "24.9.0",
+    "jest-playback": "2.0.2",
+    "prettier": "1.17.0",
+    "tslint": "5.19.0",
+    "tslint-plugin-prettier": "2.0.1",
+    "typescript": "3.6.2"
   },
   "engines": {
-    "node": ">=18"
-  },
-  "packageManager": "pnpm@8.6.6"
+    "node": ">= 12"
+  }
 }
