+++ scripts/generate.test.js
@@ -1,0 +1,1370 @@
+require("jest-playback").setup(__dirname);
+
+const generate = require("./generate");
+
+test("emoji-cheat-sheet", async () => {
+  expect(await generate()).toMatchInlineSnapshot(`
+    "# emoji-cheat-sheet
+
+    [![Up to Date](https://github.com/ikatyang/emoji-cheat-sheet/workflows/Up%20to%20Date/badge.svg)](https://github.com/ikatyang/emoji-cheat-sheet/actions?query=workflow%3A%22Up+to+Date%22)
+
+    This cheat sheet is automatically generated from [GitHub Emoji API](https://api.github.com/emojis) and [Unicode Full Emoji List](https://unicode.org/emoji/charts/full-emoji-list.html).
+
+    ## Table of Contents
+
+    - [Smileys & Emotion](#smileys--emotion)
+    - [People & Body](#people--body)
+    - [Animals & Nature](#animals--nature)
+    - [Food & Drink](#food--drink)
+    - [Travel & Places](#travel--places)
+    - [Activities](#activities)
+    - [Objects](#objects)
+    - [Symbols](#symbols)
+    - [Flags](#flags)
+    - [GitHub Custom Emoji](#github-custom-emoji)
+
+    ### Smileys & Emotion
+
+    - [Face Smiling](#face-smiling)
+    - [Face Affection](#face-affection)
+    - [Face Tongue](#face-tongue)
+    - [Face Hand](#face-hand)
+    - [Face Neutral Skeptical](#face-neutral-skeptical)
+    - [Face Sleepy](#face-sleepy)
+    - [Face Unwell](#face-unwell)
+    - [Face Hat](#face-hat)
+    - [Face Glasses](#face-glasses)
+    - [Face Concerned](#face-concerned)
+    - [Face Negative](#face-negative)
+    - [Face Costume](#face-costume)
+    - [Cat Face](#cat-face)
+    - [Monkey Face](#monkey-face)
+    - [Emotion](#emotion)
+
+    #### Face Smiling
+
+    | | ico | shortcode | ico | shortcode | |
+    | - | :-: | - | :-: | - | - |
+    | [top](#smileys--emotion) | :grinning: | \`:grinning:\` | :smiley: | \`:smiley:\` | [top](#table-of-contents) |
+    | [top](#smileys--emotion) | :smile: | \`:smile:\` | :grin: | \`:grin:\` | [top](#table-of-contents) |
+    | [top](#smileys--emotion) | :laughing: | \`:laughing:\` <br /> \`:satisfied:\` | :sweat_smile: | \`:sweat_smile:\` | [top](#table-of-contents) |
+    | [top](#smileys--emotion) | :rofl: | \`:rofl:\` | :joy: | \`:joy:\` | [top](#table-of-contents) |
+    | [top](#smileys--emotion) | :slightly_smiling_face: | \`:slightly_smiling_face:\` | :upside_down_face: | \`:upside_down_face:\` | [top](#table-of-contents) |
+    | [top](#smileys--emotion) | :wink: | \`:wink:\` | :blush: | \`:blush:\` | [top](#table-of-contents) |
+    | [top](#smileys--emotion) | :innocent: | \`:innocent:\` | | | [top](#table-of-contents) |
+
+    #### Face Affection
+
+    | | ico | shortcode | ico | shortcode | |
+    | - | :-: | - | :-: | - | - |
+    | [top](#smileys--emotion) | :heart_eyes: | \`:heart_eyes:\` | :kissing_heart: | \`:kissing_heart:\` | [top](#table-of-contents) |
+    | [top](#smileys--emotion) | :kissing: | \`:kissing:\` | :relaxed: | \`:relaxed:\` | [top](#table-of-contents) |
+    | [top](#smileys--emotion) | :kissing_closed_eyes: | \`:kissing_closed_eyes:\` | :kissing_smiling_eyes: | \`:kissing_smiling_eyes:\` | [top](#table-of-contents) |
+
+    #### Face Tongue
+
+    | | ico | shortcode | ico | shortcode | |
+    | - | :-: | - | :-: | - | - |
+    | [top](#smileys--emotion) | :yum: | \`:yum:\` | :stuck_out_tongue: | \`:stuck_out_tongue:\` | [top](#table-of-contents) |
+    | [top](#smileys--emotion) | :stuck_out_tongue_winking_eye: | \`:stuck_out_tongue_winking_eye:\` | :stuck_out_tongue_closed_eyes: | \`:stuck_out_tongue_closed_eyes:\` | [top](#table-of-contents) |
+    | [top](#smileys--emotion) | :money_mouth_face: | \`:money_mouth_face:\` | | | [top](#table-of-contents) |
+
+    #### Face Hand
+
+    | | ico | shortcode | ico | shortcode | |
+    | - | :-: | - | :-: | - | - |
+    | [top](#smileys--emotion) | :hugs: | \`:hugs:\` | :thinking: | \`:thinking:\` | [top](#table-of-contents) |
+
+    #### Face Neutral Skeptical
+
+    | | ico | shortcode | ico | shortcode | |
+    | - | :-: | - | :-: | - | - |
+    | [top](#smileys--emotion) | :zipper_mouth_face: | \`:zipper_mouth_face:\` | :neutral_face: | \`:neutral_face:\` | [top](#table-of-contents) |
+    | [top](#smileys--emotion) | :expressionless: | \`:expressionless:\` | :no_mouth: | \`:no_mouth:\` | [top](#table-of-contents) |
+    | [top](#smileys--emotion) | :smirk: | \`:smirk:\` | :unamused: | \`:unamused:\` | [top](#table-of-contents) |
+    | [top](#smileys--emotion) | :roll_eyes: | \`:roll_eyes:\` | :grimacing: | \`:grimacing:\` | [top](#table-of-contents) |
+    | [top](#smileys--emotion) | :lying_face: | \`:lying_face:\` | | | [top](#table-of-contents) |
+
+    #### Face Sleepy
+
+    | | ico | shortcode | ico | shortcode | |
+    | - | :-: | - | :-: | - | - |
+    | [top](#smileys--emotion) | :relieved: | \`:relieved:\` | :pensive: | \`:pensive:\` | [top](#table-of-contents) |
+    | [top](#smileys--emotion) | :sleepy: | \`:sleepy:\` | :drooling_face: | \`:drooling_face:\` | [top](#table-of-contents) |
+    | [top](#smileys--emotion) | :sleeping: | \`:sleeping:\` | | | [top](#table-of-contents) |
+
+    #### Face Unwell
+
+    | | ico | shortcode | ico | shortcode | |
+    | - | :-: | - | :-: | - | - |
+    | [top](#smileys--emotion) | :mask: | \`:mask:\` | :face_with_thermometer: | \`:face_with_thermometer:\` | [top](#table-of-contents) |
+    | [top](#smileys--emotion) | :face_with_head_bandage: | \`:face_with_head_bandage:\` | :nauseated_face: | \`:nauseated_face:\` | [top](#table-of-contents) |
+    | [top](#smileys--emotion) | :sneezing_face: | \`:sneezing_face:\` | :dizzy_face: | \`:dizzy_face:\` | [top](#table-of-contents) |
+
+    #### Face Hat
+
+    | | ico | shortcode | |
+    | - | :-: | - | - |
+    | [top](#smileys--emotion) | :cowboy_hat_face: | \`:cowboy_hat_face:\` | [top](#table-of-contents) |
+
+    #### Face Glasses
+
+    | | ico | shortcode | ico | shortcode | |
+    | - | :-: | - | :-: | - | - |
+    | [top](#smileys--emotion) | :sunglasses: | \`:sunglasses:\` | :nerd_face: | \`:nerd_face:\` | [top](#table-of-contents) |
+
+    #### Face Concerned
+
+    | | ico | shortcode | ico | shortcode | |
+    | - | :-: | - | :-: | - | - |
+    | [top](#smileys--emotion) | :confused: | \`:confused:\` | :worried: | \`:worried:\` | [top](#table-of-contents) |
+    | [top](#smileys--emotion) | :slightly_frowning_face: | \`:slightly_frowning_face:\` | :frowning_face: | \`:frowning_face:\` | [top](#table-of-contents) |
+    | [top](#smileys--emotion) | :open_mouth: | \`:open_mouth:\` | :hushed: | \`:hushed:\` | [top](#table-of-contents) |
+    | [top](#smileys--emotion) | :astonished: | \`:astonished:\` | :flushed: | \`:flushed:\` | [top](#table-of-contents) |
+    | [top](#smileys--emotion) | :frowning: | \`:frowning:\` | :anguished: | \`:anguished:\` | [top](#table-of-contents) |
+    | [top](#smileys--emotion) | :fearful: | \`:fearful:\` | :cold_sweat: | \`:cold_sweat:\` | [top](#table-of-contents) |
+    | [top](#smileys--emotion) | :disappointed_relieved: | \`:disappointed_relieved:\` | :cry: | \`:cry:\` | [top](#table-of-contents) |
+    | [top](#smileys--emotion) | :sob: | \`:sob:\` | :scream: | \`:scream:\` | [top](#table-of-contents) |
+    | [top](#smileys--emotion) | :confounded: | \`:confounded:\` | :persevere: | \`:persevere:\` | [top](#table-of-contents) |
+    | [top](#smileys--emotion) | :disappointed: | \`:disappointed:\` | :sweat: | \`:sweat:\` | [top](#table-of-contents) |
+    | [top](#smileys--emotion) | :weary: | \`:weary:\` | :tired_face: | \`:tired_face:\` | [top](#table-of-contents) |
+
+    #### Face Negative
+
+    | | ico | shortcode | ico | shortcode | |
+    | - | :-: | - | :-: | - | - |
+    | [top](#smileys--emotion) | :triumph: | \`:triumph:\` | :pout: | \`:pout:\` <br /> \`:rage:\` | [top](#table-of-contents) |
+    | [top](#smileys--emotion) | :angry: | \`:angry:\` | :smiling_imp: | \`:smiling_imp:\` | [top](#table-of-contents) |
+    | [top](#smileys--emotion) | :imp: | \`:imp:\` | :skull: | \`:skull:\` | [top](#table-of-contents) |
+    | [top](#smileys--emotion) | :skull_and_crossbones: | \`:skull_and_crossbones:\` | | | [top](#table-of-contents) |
+
+    #### Face Costume
+
+    | | ico | shortcode | ico | shortcode | |
+    | - | :-: | - | :-: | - | - |
+    | [top](#smileys--emotion) | :hankey: | \`:hankey:\` <br /> \`:poop:\` <br /> \`:shit:\` | :clown_face: | \`:clown_face:\` | [top](#table-of-contents) |
+    | [top](#smileys--emotion) | :japanese_ogre: | \`:japanese_ogre:\` | :japanese_goblin: | \`:japanese_goblin:\` | [top](#table-of-contents) |
+    | [top](#smileys--emotion) | :ghost: | \`:ghost:\` | :alien: | \`:alien:\` | [top](#table-of-contents) |
+    | [top](#smileys--emotion) | :space_invader: | \`:space_invader:\` | :robot: | \`:robot:\` | [top](#table-of-contents) |
+
+    #### Cat Face
+
+    | | ico | shortcode | ico | shortcode | |
+    | - | :-: | - | :-: | - | - |
+    | [top](#smileys--emotion) | :smiley_cat: | \`:smiley_cat:\` | :smile_cat: | \`:smile_cat:\` | [top](#table-of-contents) |
+    | [top](#smileys--emotion) | :joy_cat: | \`:joy_cat:\` | :heart_eyes_cat: | \`:heart_eyes_cat:\` | [top](#table-of-contents) |
+    | [top](#smileys--emotion) | :smirk_cat: | \`:smirk_cat:\` | :kissing_cat: | \`:kissing_cat:\` | [top](#table-of-contents) |
+    | [top](#smileys--emotion) | :scream_cat: | \`:scream_cat:\` | :crying_cat_face: | \`:crying_cat_face:\` | [top](#table-of-contents) |
+    | [top](#smileys--emotion) | :pouting_cat: | \`:pouting_cat:\` | | | [top](#table-of-contents) |
+
+    #### Monkey Face
+
+    | | ico | shortcode | ico | shortcode | |
+    | - | :-: | - | :-: | - | - |
+    | [top](#smileys--emotion) | :see_no_evil: | \`:see_no_evil:\` | :hear_no_evil: | \`:hear_no_evil:\` | [top](#table-of-contents) |
+    | [top](#smileys--emotion) | :speak_no_evil: | \`:speak_no_evil:\` | | | [top](#table-of-contents) |
+
+    #### Emotion
+
+    | | ico | shortcode | ico | shortcode | |
+    | - | :-: | - | :-: | - | - |
+    | [top](#smileys--emotion) | :kiss: | \`:kiss:\` | :love_letter: | \`:love_letter:\` | [top](#table-of-contents) |
+    | [top](#smileys--emotion) | :cupid: | \`:cupid:\` | :gift_heart: | \`:gift_heart:\` | [top](#table-of-contents) |
+    | [top](#smileys--emotion) | :sparkling_heart: | \`:sparkling_heart:\` | :heartpulse: | \`:heartpulse:\` | [top](#table-of-contents) |
+    | [top](#smileys--emotion) | :heartbeat: | \`:heartbeat:\` | :revolving_hearts: | \`:revolving_hearts:\` | [top](#table-of-contents) |
+    | [top](#smileys--emotion) | :two_hearts: | \`:two_hearts:\` | :heart_decoration: | \`:heart_decoration:\` | [top](#table-of-contents) |
+    | [top](#smileys--emotion) | :heavy_heart_exclamation: | \`:heavy_heart_exclamation:\` | :broken_heart: | \`:broken_heart:\` | [top](#table-of-contents) |
+    | [top](#smileys--emotion) | :heart: | \`:heart:\` | :yellow_heart: | \`:yellow_heart:\` | [top](#table-of-contents) |
+    | [top](#smileys--emotion) | :green_heart: | \`:green_heart:\` | :blue_heart: | \`:blue_heart:\` | [top](#table-of-contents) |
+    | [top](#smileys--emotion) | :purple_heart: | \`:purple_heart:\` | :black_heart: | \`:black_heart:\` | [top](#table-of-contents) |
+    | [top](#smileys--emotion) | :100: | \`:100:\` | :anger: | \`:anger:\` | [top](#table-of-contents) |
+    | [top](#smileys--emotion) | :boom: | \`:boom:\` <br /> \`:collision:\` | :dizzy: | \`:dizzy:\` | [top](#table-of-contents) |
+    | [top](#smileys--emotion) | :sweat_drops: | \`:sweat_drops:\` | :dash: | \`:dash:\` | [top](#table-of-contents) |
+    | [top](#smileys--emotion) | :hole: | \`:hole:\` | :bomb: | \`:bomb:\` | [top](#table-of-contents) |
+    | [top](#smileys--emotion) | :speech_balloon: | \`:speech_balloon:\` | :eye_speech_bubble: | \`:eye_speech_bubble:\` | [top](#table-of-contents) |
+    | [top](#smileys--emotion) | :right_anger_bubble: | \`:right_anger_bubble:\` | :thought_balloon: | \`:thought_balloon:\` | [top](#table-of-contents) |
+    | [top](#smileys--emotion) | :zzz: | \`:zzz:\` | | | [top](#table-of-contents) |
+
+    ### People & Body
+
+    - [Hand Fingers Open](#hand-fingers-open)
+    - [Hand Fingers Partial](#hand-fingers-partial)
+    - [Hand Single Finger](#hand-single-finger)
+    - [Hand Fingers Closed](#hand-fingers-closed)
+    - [Hands](#hands)
+    - [Hand Prop](#hand-prop)
+    - [Body Parts](#body-parts)
+    - [Person](#person)
+    - [Person Gesture](#person-gesture)
+    - [Person Role](#person-role)
+    - [Person Fantasy](#person-fantasy)
+    - [Person Activity](#person-activity)
+    - [Person Sport](#person-sport)
+    - [Person Resting](#person-resting)
+    - [Family](#family)
+    - [Person Symbol](#person-symbol)
+
+    #### Hand Fingers Open
+
+    | | ico | shortcode | ico | shortcode | |
+    | - | :-: | - | :-: | - | - |
+    | [top](#people--body) | :wave: | \`:wave:\` | :raised_back_of_hand: | \`:raised_back_of_hand:\` | [top](#table-of-contents) |
+    | [top](#people--body) | :raised_hand_with_fingers_splayed: | \`:raised_hand_with_fingers_splayed:\` | :hand: | \`:hand:\` <br /> \`:raised_hand:\` | [top](#table-of-contents) |
+    | [top](#people--body) | :vulcan_salute: | \`:vulcan_salute:\` | | | [top](#table-of-contents) |
+
+    #### Hand Fingers Partial
+
+    | | ico | shortcode | ico | shortcode | |
+    | - | :-: | - | :-: | - | - |
+    | [top](#people--body) | :ok_hand: | \`:ok_hand:\` | :v: | \`:v:\` | [top](#table-of-contents) |
+    | [top](#people--body) | :crossed_fingers: | \`:crossed_fingers:\` | :metal: | \`:metal:\` | [top](#table-of-contents) |
+    | [top](#people--body) | :call_me_hand: | \`:call_me_hand:\` | | | [top](#table-of-contents) |
+
+    #### Hand Single Finger
+
+    | | ico | shortcode | ico | shortcode | |
+    | - | :-: | - | :-: | - | - |
+    | [top](#people--body) | :point_left: | \`:point_left:\` | :point_right: | \`:point_right:\` | [top](#table-of-contents) |
+    | [top](#people--body) | :point_up_2: | \`:point_up_2:\` | :fu: | \`:fu:\` <br /> \`:middle_finger:\` | [top](#table-of-contents) |
+    | [top](#people--body) | :point_down: | \`:point_down:\` | :point_up: | \`:point_up:\` | [top](#table-of-contents) |
+
+    #### Hand Fingers Closed
+
+    | | ico | shortcode | ico | shortcode | |
+    | - | :-: | - | :-: | - | - |
+    | [top](#people--body) | :+1: | \`:+1:\` <br /> \`:thumbsup:\` | :-1: | \`:-1:\` <br /> \`:thumbsdown:\` | [top](#table-of-contents) |
+    | [top](#people--body) | :fist: | \`:fist:\` <br /> \`:fist_raised:\` | :facepunch: | \`:facepunch:\` <br /> \`:fist_oncoming:\` <br /> \`:punch:\` | [top](#table-of-contents) |
+    | [top](#people--body) | :fist_left: | \`:fist_left:\` | :fist_right: | \`:fist_right:\` | [top](#table-of-contents) |
+
+    #### Hands
+
+    | | ico | shortcode | ico | shortcode | |
+    | - | :-: | - | :-: | - | - |
+    | [top](#people--body) | :clap: | \`:clap:\` | :raised_hands: | \`:raised_hands:\` | [top](#table-of-contents) |
+    | [top](#people--body) | :open_hands: | \`:open_hands:\` | :handshake: | \`:handshake:\` | [top](#table-of-contents) |
+    | [top](#people--body) | :pray: | \`:pray:\` | | | [top](#table-of-contents) |
+
+    #### Hand Prop
+
+    | | ico | shortcode | ico | shortcode | |
+    | - | :-: | - | :-: | - | - |
+    | [top](#people--body) | :writing_hand: | \`:writing_hand:\` | :nail_care: | \`:nail_care:\` | [top](#table-of-contents) |
+    | [top](#people--body) | :selfie: | \`:selfie:\` | | | [top](#table-of-contents) |
+
+    #### Body Parts
+
+    | | ico | shortcode | ico | shortcode | |
+    | - | :-: | - | :-: | - | - |
+    | [top](#people--body) | :muscle: | \`:muscle:\` | :ear: | \`:ear:\` | [top](#table-of-contents) |
+    | [top](#people--body) | :nose: | \`:nose:\` | :eyes: | \`:eyes:\` | [top](#table-of-contents) |
+    | [top](#people--body) | :eye: | \`:eye:\` | :tongue: | \`:tongue:\` | [top](#table-of-contents) |
+    | [top](#people--body) | :lips: | \`:lips:\` | | | [top](#table-of-contents) |
+
+    #### Person
+
+    | | ico | shortcode | ico | shortcode | |
+    | - | :-: | - | :-: | - | - |
+    | [top](#people--body) | :baby: | \`:baby:\` | :boy: | \`:boy:\` | [top](#table-of-contents) |
+    | [top](#people--body) | :girl: | \`:girl:\` | :blonde_man: | \`:blonde_man:\` <br /> \`:person_with_blond_hair:\` | [top](#table-of-contents) |
+    | [top](#people--body) | :man: | \`:man:\` | :woman: | \`:woman:\` | [top](#table-of-contents) |
+    | [top](#people--body) | :blonde_woman: | \`:blonde_woman:\` | :older_man: | \`:older_man:\` | [top](#table-of-contents) |
+    | [top](#people--body) | :older_woman: | \`:older_woman:\` | | | [top](#table-of-contents) |
+
+    #### Person Gesture
+
+    | | ico | shortcode | ico | shortcode | |
+    | - | :-: | - | :-: | - | - |
+    | [top](#people--body) | :frowning_woman: | \`:frowning_woman:\` <br /> \`:person_frowning:\` | :frowning_man: | \`:frowning_man:\` | [top](#table-of-contents) |
+    | [top](#people--body) | :person_with_pouting_face: | \`:person_with_pouting_face:\` <br /> \`:pouting_woman:\` | :pouting_man: | \`:pouting_man:\` | [top](#table-of-contents) |
+    | [top](#people--body) | :ng_woman: | \`:ng_woman:\` <br /> \`:no_good:\` <br /> \`:no_good_woman:\` | :ng_man: | \`:ng_man:\` <br /> \`:no_good_man:\` | [top](#table-of-contents) |
+    | [top](#people--body) | :ok_woman: | \`:ok_woman:\` | :ok_man: | \`:ok_man:\` | [top](#table-of-contents) |
+    | [top](#people--body) | :information_desk_person: | \`:information_desk_person:\` <br /> \`:sassy_woman:\` <br /> \`:tipping_hand_woman:\` | :sassy_man: | \`:sassy_man:\` <br /> \`:tipping_hand_man:\` | [top](#table-of-contents) |
+    | [top](#people--body) | :raising_hand: | \`:raising_hand:\` <br /> \`:raising_hand_woman:\` | :raising_hand_man: | \`:raising_hand_man:\` | [top](#table-of-contents) |
+    | [top](#people--body) | :bow: | \`:bow:\` <br /> \`:bowing_man:\` | :bowing_woman: | \`:bowing_woman:\` | [top](#table-of-contents) |
+    | [top](#people--body) | :man_facepalming: | \`:man_facepalming:\` | :woman_facepalming: | \`:woman_facepalming:\` | [top](#table-of-contents) |
+    | [top](#people--body) | :man_shrugging: | \`:man_shrugging:\` | :woman_shrugging: | \`:woman_shrugging:\` | [top](#table-of-contents) |
+
+    #### Person Role
+
+    | | ico | shortcode | ico | shortcode | |
+    | - | :-: | - | :-: | - | - |
+    | [top](#people--body) | :man_health_worker: | \`:man_health_worker:\` | :woman_health_worker: | \`:woman_health_worker:\` | [top](#table-of-contents) |
+    | [top](#people--body) | :man_student: | \`:man_student:\` | :woman_student: | \`:woman_student:\` | [top](#table-of-contents) |
+    | [top](#people--body) | :man_teacher: | \`:man_teacher:\` | :woman_teacher: | \`:woman_teacher:\` | [top](#table-of-contents) |
+    | [top](#people--body) | :man_judge: | \`:man_judge:\` | :woman_judge: | \`:woman_judge:\` | [top](#table-of-contents) |
+    | [top](#people--body) | :man_farmer: | \`:man_farmer:\` | :woman_farmer: | \`:woman_farmer:\` | [top](#table-of-contents) |
+    | [top](#people--body) | :man_cook: | \`:man_cook:\` | :woman_cook: | \`:woman_cook:\` | [top](#table-of-contents) |
+    | [top](#people--body) | :man_mechanic: | \`:man_mechanic:\` | :woman_mechanic: | \`:woman_mechanic:\` | [top](#table-of-contents) |
+    | [top](#people--body) | :man_factory_worker: | \`:man_factory_worker:\` | :woman_factory_worker: | \`:woman_factory_worker:\` | [top](#table-of-contents) |
+    | [top](#people--body) | :man_office_worker: | \`:man_office_worker:\` | :woman_office_worker: | \`:woman_office_worker:\` | [top](#table-of-contents) |
+    | [top](#people--body) | :man_scientist: | \`:man_scientist:\` | :woman_scientist: | \`:woman_scientist:\` | [top](#table-of-contents) |
+    | [top](#people--body) | :man_technologist: | \`:man_technologist:\` | :woman_technologist: | \`:woman_technologist:\` | [top](#table-of-contents) |
+    | [top](#people--body) | :man_singer: | \`:man_singer:\` | :woman_singer: | \`:woman_singer:\` | [top](#table-of-contents) |
+    | [top](#people--body) | :man_artist: | \`:man_artist:\` | :woman_artist: | \`:woman_artist:\` | [top](#table-of-contents) |
+    | [top](#people--body) | :man_pilot: | \`:man_pilot:\` | :woman_pilot: | \`:woman_pilot:\` | [top](#table-of-contents) |
+    | [top](#people--body) | :man_astronaut: | \`:man_astronaut:\` | :woman_astronaut: | \`:woman_astronaut:\` | [top](#table-of-contents) |
+    | [top](#people--body) | :man_firefighter: | \`:man_firefighter:\` | :woman_firefighter: | \`:woman_firefighter:\` | [top](#table-of-contents) |
+    | [top](#people--body) | :cop: | \`:cop:\` <br /> \`:policeman:\` | :policewoman: | \`:policewoman:\` | [top](#table-of-contents) |
+    | [top](#people--body) | :detective: | \`:detective:\` <br /> \`:male_detective:\` | :female_detective: | \`:female_detective:\` | [top](#table-of-contents) |
+    | [top](#people--body) | :guardsman: | \`:guardsman:\` | :guardswoman: | \`:guardswoman:\` | [top](#table-of-contents) |
+    | [top](#people--body) | :construction_worker: | \`:construction_worker:\` <br /> \`:construction_worker_man:\` | :construction_worker_woman: | \`:construction_worker_woman:\` | [top](#table-of-contents) |
+    | [top](#people--body) | :prince: | \`:prince:\` | :princess: | \`:princess:\` | [top](#table-of-contents) |
+    | [top](#people--body) | :man_with_turban: | \`:man_with_turban:\` | :woman_with_turban: | \`:woman_with_turban:\` | [top](#table-of-contents) |
+    | [top](#people--body) | :man_with_gua_pi_mao: | \`:man_with_gua_pi_mao:\` | :man_in_tuxedo: | \`:man_in_tuxedo:\` | [top](#table-of-contents) |
+    | [top](#people--body) | :bride_with_veil: | \`:bride_with_veil:\` | :pregnant_woman: | \`:pregnant_woman:\` | [top](#table-of-contents) |
+
+    #### Person Fantasy
+
+    | | ico | shortcode | ico | shortcode | |
+    | - | :-: | - | :-: | - | - |
+    | [top](#people--body) | :angel: | \`:angel:\` | :santa: | \`:santa:\` | [top](#table-of-contents) |
+    | [top](#people--body) | :mrs_claus: | \`:mrs_claus:\` | | | [top](#table-of-contents) |
+
+    #### Person Activity
+
+    | | ico | shortcode | ico | shortcode | |
+    | - | :-: | - | :-: | - | - |
+    | [top](#people--body) | :massage: | \`:massage:\` <br /> \`:massage_woman:\` | :massage_man: | \`:massage_man:\` | [top](#table-of-contents) |
+    | [top](#people--body) | :haircut: | \`:haircut:\` <br /> \`:haircut_woman:\` | :haircut_man: | \`:haircut_man:\` | [top](#table-of-contents) |
+    | [top](#people--body) | :walking: | \`:walking:\` <br /> \`:walking_man:\` | :walking_woman: | \`:walking_woman:\` | [top](#table-of-contents) |
+    | [top](#people--body) | :runner: | \`:runner:\` <br /> \`:running:\` <br /> \`:running_man:\` | :running_woman: | \`:running_woman:\` | [top](#table-of-contents) |
+    | [top](#people--body) | :dancer: | \`:dancer:\` | :man_dancing: | \`:man_dancing:\` | [top](#table-of-contents) |
+    | [top](#people--body) | :business_suit_levitating: | \`:business_suit_levitating:\` | :dancers: | \`:dancers:\` <br /> \`:dancing_women:\` | [top](#table-of-contents) |
+    | [top](#people--body) | :dancing_men: | \`:dancing_men:\` | | | [top](#table-of-contents) |
+
+    #### Person Sport
+
+    | | ico | shortcode | ico | shortcode | |
+    | - | :-: | - | :-: | - | - |
+    | [top](#people--body) | :person_fencing: | \`:person_fencing:\` | :horse_racing: | \`:horse_racing:\` | [top](#table-of-contents) |
+    | [top](#people--body) | :skier: | \`:skier:\` | :snowboarder: | \`:snowboarder:\` | [top](#table-of-contents) |
+    | [top](#people--body) | :golfing_man: | \`:golfing_man:\` | :golfing_woman: | \`:golfing_woman:\` | [top](#table-of-contents) |
+    | [top](#people--body) | :surfer: | \`:surfer:\` <br /> \`:surfing_man:\` | :surfing_woman: | \`:surfing_woman:\` | [top](#table-of-contents) |
+    | [top](#people--body) | :rowboat: | \`:rowboat:\` <br /> \`:rowing_man:\` | :rowing_woman: | \`:rowing_woman:\` | [top](#table-of-contents) |
+    | [top](#people--body) | :swimmer: | \`:swimmer:\` <br /> \`:swimming_man:\` | :swimming_woman: | \`:swimming_woman:\` | [top](#table-of-contents) |
+    | [top](#people--body) | :basketball_man: | \`:basketball_man:\` | :basketball_woman: | \`:basketball_woman:\` | [top](#table-of-contents) |
+    | [top](#people--body) | :weight_lifting_man: | \`:weight_lifting_man:\` | :weight_lifting_woman: | \`:weight_lifting_woman:\` | [top](#table-of-contents) |
+    | [top](#people--body) | :bicyclist: | \`:bicyclist:\` <br /> \`:biking_man:\` | :biking_woman: | \`:biking_woman:\` | [top](#table-of-contents) |
+    | [top](#people--body) | :mountain_bicyclist: | \`:mountain_bicyclist:\` <br /> \`:mountain_biking_man:\` | :mountain_biking_woman: | \`:mountain_biking_woman:\` | [top](#table-of-contents) |
+    | [top](#people--body) | :man_cartwheeling: | \`:man_cartwheeling:\` | :woman_cartwheeling: | \`:woman_cartwheeling:\` | [top](#table-of-contents) |
+    | [top](#people--body) | :men_wrestling: | \`:men_wrestling:\` | :women_wrestling: | \`:women_wrestling:\` | [top](#table-of-contents) |
+    | [top](#people--body) | :man_playing_water_polo: | \`:man_playing_water_polo:\` | :woman_playing_water_polo: | \`:woman_playing_water_polo:\` | [top](#table-of-contents) |
+    | [top](#people--body) | :man_playing_handball: | \`:man_playing_handball:\` | :woman_playing_handball: | \`:woman_playing_handball:\` | [top](#table-of-contents) |
+    | [top](#people--body) | :man_juggling: | \`:man_juggling:\` | :woman_juggling: | \`:woman_juggling:\` | [top](#table-of-contents) |
+
+    #### Person Resting
+
+    | | ico | shortcode | ico | shortcode | |
+    | - | :-: | - | :-: | - | - |
+    | [top](#people--body) | :bath: | \`:bath:\` | :sleeping_bed: | \`:sleeping_bed:\` | [top](#table-of-contents) |
+
+    #### Family
+
+    | | ico | shortcode | ico | shortcode | |
+    | - | :-: | - | :-: | - | - |
+    | [top](#people--body) | :two_women_holding_hands: | \`:two_women_holding_hands:\` | :couple: | \`:couple:\` | [top](#table-of-contents) |
+    | [top](#people--body) | :two_men_holding_hands: | \`:two_men_holding_hands:\` | :couplekiss_man_woman: | \`:couplekiss_man_woman:\` | [top](#table-of-contents) |
+    | [top](#people--body) | :couplekiss_man_man: | \`:couplekiss_man_man:\` | :couplekiss_woman_woman: | \`:couplekiss_woman_woman:\` | [top](#table-of-contents) |
+    | [top](#people--body) | :couple_with_heart: | \`:couple_with_heart:\` <br /> \`:couple_with_heart_woman_man:\` | :couple_with_heart_man_man: | \`:couple_with_heart_man_man:\` | [top](#table-of-contents) |
+    | [top](#people--body) | :couple_with_heart_woman_woman: | \`:couple_with_heart_woman_woman:\` | :family: | \`:family:\` <br /> \`:family_man_woman_boy:\` | [top](#table-of-contents) |
+    | [top](#people--body) | :family_man_woman_girl: | \`:family_man_woman_girl:\` | :family_man_woman_girl_boy: | \`:family_man_woman_girl_boy:\` | [top](#table-of-contents) |
+    | [top](#people--body) | :family_man_woman_boy_boy: | \`:family_man_woman_boy_boy:\` | :family_man_woman_girl_girl: | \`:family_man_woman_girl_girl:\` | [top](#table-of-contents) |
+    | [top](#people--body) | :family_man_man_boy: | \`:family_man_man_boy:\` | :family_man_man_girl: | \`:family_man_man_girl:\` | [top](#table-of-contents) |
+    | [top](#people--body) | :family_man_man_girl_boy: | \`:family_man_man_girl_boy:\` | :family_man_man_boy_boy: | \`:family_man_man_boy_boy:\` | [top](#table-of-contents) |
+    | [top](#people--body) | :family_man_man_girl_girl: | \`:family_man_man_girl_girl:\` | :family_woman_woman_boy: | \`:family_woman_woman_boy:\` | [top](#table-of-contents) |
+    | [top](#people--body) | :family_woman_woman_girl: | \`:family_woman_woman_girl:\` | :family_woman_woman_girl_boy: | \`:family_woman_woman_girl_boy:\` | [top](#table-of-contents) |
+    | [top](#people--body) | :family_woman_woman_boy_boy: | \`:family_woman_woman_boy_boy:\` | :family_woman_woman_girl_girl: | \`:family_woman_woman_girl_girl:\` | [top](#table-of-contents) |
+    | [top](#people--body) | :family_man_boy: | \`:family_man_boy:\` | :family_man_boy_boy: | \`:family_man_boy_boy:\` | [top](#table-of-contents) |
+    | [top](#people--body) | :family_man_girl: | \`:family_man_girl:\` | :family_man_girl_boy: | \`:family_man_girl_boy:\` | [top](#table-of-contents) |
+    | [top](#people--body) | :family_man_girl_girl: | \`:family_man_girl_girl:\` | :family_woman_boy: | \`:family_woman_boy:\` | [top](#table-of-contents) |
+    | [top](#people--body) | :family_woman_boy_boy: | \`:family_woman_boy_boy:\` | :family_woman_girl: | \`:family_woman_girl:\` | [top](#table-of-contents) |
+    | [top](#people--body) | :family_woman_girl_boy: | \`:family_woman_girl_boy:\` | :family_woman_girl_girl: | \`:family_woman_girl_girl:\` | [top](#table-of-contents) |
+
+    #### Person Symbol
+
+    | | ico | shortcode | ico | shortcode | |
+    | - | :-: | - | :-: | - | - |
+    | [top](#people--body) | :speaking_head: | \`:speaking_head:\` | :bust_in_silhouette: | \`:bust_in_silhouette:\` | [top](#table-of-contents) |
+    | [top](#people--body) | :busts_in_silhouette: | \`:busts_in_silhouette:\` | :footprints: | \`:footprints:\` | [top](#table-of-contents) |
+
+    ### Animals & Nature
+
+    - [Animal Mammal](#animal-mammal)
+    - [Animal Bird](#animal-bird)
+    - [Animal Amphibian](#animal-amphibian)
+    - [Animal Reptile](#animal-reptile)
+    - [Animal Marine](#animal-marine)
+    - [Animal Bug](#animal-bug)
+    - [Plant Flower](#plant-flower)
+    - [Plant Other](#plant-other)
+
+    #### Animal Mammal
+
+    | | ico | shortcode | ico | shortcode | |
+    | - | :-: | - | :-: | - | - |
+    | [top](#animals--nature) | :monkey_face: | \`:monkey_face:\` | :monkey: | \`:monkey:\` | [top](#table-of-contents) |
+    | [top](#animals--nature) | :gorilla: | \`:gorilla:\` | :dog: | \`:dog:\` | [top](#table-of-contents) |
+    | [top](#animals--nature) | :dog2: | \`:dog2:\` | :poodle: | \`:poodle:\` | [top](#table-of-contents) |
+    | [top](#animals--nature) | :wolf: | \`:wolf:\` | :fox_face: | \`:fox_face:\` | [top](#table-of-contents) |
+    | [top](#animals--nature) | :cat: | \`:cat:\` | :cat2: | \`:cat2:\` | [top](#table-of-contents) |
+    | [top](#animals--nature) | :lion: | \`:lion:\` | :tiger: | \`:tiger:\` | [top](#table-of-contents) |
+    | [top](#animals--nature) | :tiger2: | \`:tiger2:\` | :leopard: | \`:leopard:\` | [top](#table-of-contents) |
+    | [top](#animals--nature) | :horse: | \`:horse:\` | :racehorse: | \`:racehorse:\` | [top](#table-of-contents) |
+    | [top](#animals--nature) | :unicorn: | \`:unicorn:\` | :deer: | \`:deer:\` | [top](#table-of-contents) |
+    | [top](#animals--nature) | :cow: | \`:cow:\` | :ox: | \`:ox:\` | [top](#table-of-contents) |
+    | [top](#animals--nature) | :water_buffalo: | \`:water_buffalo:\` | :cow2: | \`:cow2:\` | [top](#table-of-contents) |
+    | [top](#animals--nature) | :pig: | \`:pig:\` | :pig2: | \`:pig2:\` | [top](#table-of-contents) |
+    | [top](#animals--nature) | :boar: | \`:boar:\` | :pig_nose: | \`:pig_nose:\` | [top](#table-of-contents) |
+    | [top](#animals--nature) | :ram: | \`:ram:\` | :sheep: | \`:sheep:\` | [top](#table-of-contents) |
+    | [top](#animals--nature) | :goat: | \`:goat:\` | :dromedary_camel: | \`:dromedary_camel:\` | [top](#table-of-contents) |
+    | [top](#animals--nature) | :camel: | \`:camel:\` | :elephant: | \`:elephant:\` | [top](#table-of-contents) |
+    | [top](#animals--nature) | :rhinoceros: | \`:rhinoceros:\` | :mouse: | \`:mouse:\` | [top](#table-of-contents) |
+    | [top](#animals--nature) | :mouse2: | \`:mouse2:\` | :rat: | \`:rat:\` | [top](#table-of-contents) |
+    | [top](#animals--nature) | :hamster: | \`:hamster:\` | :rabbit: | \`:rabbit:\` | [top](#table-of-contents) |
+    | [top](#animals--nature) | :rabbit2: | \`:rabbit2:\` | :chipmunk: | \`:chipmunk:\` | [top](#table-of-contents) |
+    | [top](#animals--nature) | :bat: | \`:bat:\` | :bear: | \`:bear:\` | [top](#table-of-contents) |
+    | [top](#animals--nature) | :koala: | \`:koala:\` | :panda_face: | \`:panda_face:\` | [top](#table-of-contents) |
+    | [top](#animals--nature) | :feet: | \`:feet:\` <br /> \`:paw_prints:\` | | | [top](#table-of-contents) |
+
+    #### Animal Bird
+
+    | | ico | shortcode | ico | shortcode | |
+    | - | :-: | - | :-: | - | - |
+    | [top](#animals--nature) | :turkey: | \`:turkey:\` | :chicken: | \`:chicken:\` | [top](#table-of-contents) |
+    | [top](#animals--nature) | :rooster: | \`:rooster:\` | :hatching_chick: | \`:hatching_chick:\` | [top](#table-of-contents) |
+    | [top](#animals--nature) | :baby_chick: | \`:baby_chick:\` | :hatched_chick: | \`:hatched_chick:\` | [top](#table-of-contents) |
+    | [top](#animals--nature) | :bird: | \`:bird:\` | :penguin: | \`:penguin:\` | [top](#table-of-contents) |
+    | [top](#animals--nature) | :dove: | \`:dove:\` | :eagle: | \`:eagle:\` | [top](#table-of-contents) |
+    | [top](#animals--nature) | :duck: | \`:duck:\` | :owl: | \`:owl:\` | [top](#table-of-contents) |
+
+    #### Animal Amphibian
+
+    | | ico | shortcode | |
+    | - | :-: | - | - |
+    | [top](#animals--nature) | :frog: | \`:frog:\` | [top](#table-of-contents) |
+
+    #### Animal Reptile
+
+    | | ico | shortcode | ico | shortcode | |
+    | - | :-: | - | :-: | - | - |
+    | [top](#animals--nature) | :crocodile: | \`:crocodile:\` | :turtle: | \`:turtle:\` | [top](#table-of-contents) |
+    | [top](#animals--nature) | :lizard: | \`:lizard:\` | :snake: | \`:snake:\` | [top](#table-of-contents) |
+    | [top](#animals--nature) | :dragon_face: | \`:dragon_face:\` | :dragon: | \`:dragon:\` | [top](#table-of-contents) |
+
+    #### Animal Marine
+
+    | | ico | shortcode | ico | shortcode | |
+    | - | :-: | - | :-: | - | - |
+    | [top](#animals--nature) | :whale: | \`:whale:\` | :whale2: | \`:whale2:\` | [top](#table-of-contents) |
+    | [top](#animals--nature) | :dolphin: | \`:dolphin:\` <br /> \`:flipper:\` | :fish: | \`:fish:\` | [top](#table-of-contents) |
+    | [top](#animals--nature) | :tropical_fish: | \`:tropical_fish:\` | :blowfish: | \`:blowfish:\` | [top](#table-of-contents) |
+    | [top](#animals--nature) | :shark: | \`:shark:\` | :octopus: | \`:octopus:\` | [top](#table-of-contents) |
+    | [top](#animals--nature) | :shell: | \`:shell:\` | | | [top](#table-of-contents) |
+
+    #### Animal Bug
+
+    | | ico | shortcode | ico | shortcode | |
+    | - | :-: | - | :-: | - | - |
+    | [top](#animals--nature) | :snail: | \`:snail:\` | :butterfly: | \`:butterfly:\` | [top](#table-of-contents) |
+    | [top](#animals--nature) | :bug: | \`:bug:\` | :ant: | \`:ant:\` | [top](#table-of-contents) |
+    | [top](#animals--nature) | :bee: | \`:bee:\` <br /> \`:honeybee:\` | :beetle: | \`:beetle:\` | [top](#table-of-contents) |
+    | [top](#animals--nature) | :spider: | \`:spider:\` | :spider_web: | \`:spider_web:\` | [top](#table-of-contents) |
+    | [top](#animals--nature) | :scorpion: | \`:scorpion:\` | | | [top](#table-of-contents) |
+
+    #### Plant Flower
+
+    | | ico | shortcode | ico | shortcode | |
+    | - | :-: | - | :-: | - | - |
+    | [top](#animals--nature) | :bouquet: | \`:bouquet:\` | :cherry_blossom: | \`:cherry_blossom:\` | [top](#table-of-contents) |
+    | [top](#animals--nature) | :white_flower: | \`:white_flower:\` | :rosette: | \`:rosette:\` | [top](#table-of-contents) |
+    | [top](#animals--nature) | :rose: | \`:rose:\` | :wilted_flower: | \`:wilted_flower:\` | [top](#table-of-contents) |
+    | [top](#animals--nature) | :hibiscus: | \`:hibiscus:\` | :sunflower: | \`:sunflower:\` | [top](#table-of-contents) |
+    | [top](#animals--nature) | :blossom: | \`:blossom:\` | :tulip: | \`:tulip:\` | [top](#table-of-contents) |
+
+    #### Plant Other
+
+    | | ico | shortcode | ico | shortcode | |
+    | - | :-: | - | :-: | - | - |
+    | [top](#animals--nature) | :seedling: | \`:seedling:\` | :evergreen_tree: | \`:evergreen_tree:\` | [top](#table-of-contents) |
+    | [top](#animals--nature) | :deciduous_tree: | \`:deciduous_tree:\` | :palm_tree: | \`:palm_tree:\` | [top](#table-of-contents) |
+    | [top](#animals--nature) | :cactus: | \`:cactus:\` | :ear_of_rice: | \`:ear_of_rice:\` | [top](#table-of-contents) |
+    | [top](#animals--nature) | :herb: | \`:herb:\` | :shamrock: | \`:shamrock:\` | [top](#table-of-contents) |
+    | [top](#animals--nature) | :four_leaf_clover: | \`:four_leaf_clover:\` | :maple_leaf: | \`:maple_leaf:\` | [top](#table-of-contents) |
+    | [top](#animals--nature) | :fallen_leaf: | \`:fallen_leaf:\` | :leaves: | \`:leaves:\` | [top](#table-of-contents) |
+
+    ### Food & Drink
+
+    - [Food Fruit](#food-fruit)
+    - [Food Vegetable](#food-vegetable)
+    - [Food Prepared](#food-prepared)
+    - [Food Asian](#food-asian)
+    - [Food Marine](#food-marine)
+    - [Food Sweet](#food-sweet)
+    - [Drink](#drink)
+    - [Dishware](#dishware)
+
+    #### Food Fruit
+
+    | | ico | shortcode | ico | shortcode | |
+    | - | :-: | - | :-: | - | - |
+    | [top](#food--drink) | :grapes: | \`:grapes:\` | :melon: | \`:melon:\` | [top](#table-of-contents) |
+    | [top](#food--drink) | :watermelon: | \`:watermelon:\` | :mandarin: | \`:mandarin:\` <br /> \`:orange:\` <br /> \`:tangerine:\` | [top](#table-of-contents) |
+    | [top](#food--drink) | :lemon: | \`:lemon:\` | :banana: | \`:banana:\` | [top](#table-of-contents) |
+    | [top](#food--drink) | :pineapple: | \`:pineapple:\` | :apple: | \`:apple:\` | [top](#table-of-contents) |
+    | [top](#food--drink) | :green_apple: | \`:green_apple:\` | :pear: | \`:pear:\` | [top](#table-of-contents) |
+    | [top](#food--drink) | :peach: | \`:peach:\` | :cherries: | \`:cherries:\` | [top](#table-of-contents) |
+    | [top](#food--drink) | :strawberry: | \`:strawberry:\` | :kiwi_fruit: | \`:kiwi_fruit:\` | [top](#table-of-contents) |
+    | [top](#food--drink) | :tomato: | \`:tomato:\` | | | [top](#table-of-contents) |
+
+    #### Food Vegetable
+
+    | | ico | shortcode | ico | shortcode | |
+    | - | :-: | - | :-: | - | - |
+    | [top](#food--drink) | :avocado: | \`:avocado:\` | :eggplant: | \`:eggplant:\` | [top](#table-of-contents) |
+    | [top](#food--drink) | :potato: | \`:potato:\` | :carrot: | \`:carrot:\` | [top](#table-of-contents) |
+    | [top](#food--drink) | :corn: | \`:corn:\` | :hot_pepper: | \`:hot_pepper:\` | [top](#table-of-contents) |
+    | [top](#food--drink) | :cucumber: | \`:cucumber:\` | :mushroom: | \`:mushroom:\` | [top](#table-of-contents) |
+    | [top](#food--drink) | :peanuts: | \`:peanuts:\` | :chestnut: | \`:chestnut:\` | [top](#table-of-contents) |
+
+    #### Food Prepared
+
+    | | ico | shortcode | ico | shortcode | |
+    | - | :-: | - | :-: | - | - |
+    | [top](#food--drink) | :bread: | \`:bread:\` | :croissant: | \`:croissant:\` | [top](#table-of-contents) |
+    | [top](#food--drink) | :baguette_bread: | \`:baguette_bread:\` | :pancakes: | \`:pancakes:\` | [top](#table-of-contents) |
+    | [top](#food--drink) | :cheese: | \`:cheese:\` | :meat_on_bone: | \`:meat_on_bone:\` | [top](#table-of-contents) |
+    | [top](#food--drink) | :poultry_leg: | \`:poultry_leg:\` | :bacon: | \`:bacon:\` | [top](#table-of-contents) |
+    | [top](#food--drink) | :hamburger: | \`:hamburger:\` | :fries: | \`:fries:\` | [top](#table-of-contents) |
+    | [top](#food--drink) | :pizza: | \`:pizza:\` | :hotdog: | \`:hotdog:\` | [top](#table-of-contents) |
+    | [top](#food--drink) | :taco: | \`:taco:\` | :burrito: | \`:burrito:\` | [top](#table-of-contents) |
+    | [top](#food--drink) | :stuffed_flatbread: | \`:stuffed_flatbread:\` | :egg: | \`:egg:\` | [top](#table-of-contents) |
+    | [top](#food--drink) | :fried_egg: | \`:fried_egg:\` | :shallow_pan_of_food: | \`:shallow_pan_of_food:\` | [top](#table-of-contents) |
+    | [top](#food--drink) | :stew: | \`:stew:\` | :green_salad: | \`:green_salad:\` | [top](#table-of-contents) |
+    | [top](#food--drink) | :popcorn: | \`:popcorn:\` | | | [top](#table-of-contents) |
+
+    #### Food Asian
+
+    | | ico | shortcode | ico | shortcode | |
+    | - | :-: | - | :-: | - | - |
+    | [top](#food--drink) | :bento: | \`:bento:\` | :rice_cracker: | \`:rice_cracker:\` | [top](#table-of-contents) |
+    | [top](#food--drink) | :rice_ball: | \`:rice_ball:\` | :rice: | \`:rice:\` | [top](#table-of-contents) |
+    | [top](#food--drink) | :curry: | \`:curry:\` | :ramen: | \`:ramen:\` | [top](#table-of-contents) |
+    | [top](#food--drink) | :spaghetti: | \`:spaghetti:\` | :sweet_potato: | \`:sweet_potato:\` | [top](#table-of-contents) |
+    | [top](#food--drink) | :oden: | \`:oden:\` | :sushi: | \`:sushi:\` | [top](#table-of-contents) |
+    | [top](#food--drink) | :fried_shrimp: | \`:fried_shrimp:\` | :fish_cake: | \`:fish_cake:\` | [top](#table-of-contents) |
+    | [top](#food--drink) | :dango: | \`:dango:\` | | | [top](#table-of-contents) |
+
+    #### Food Marine
+
+    | | ico | shortcode | ico | shortcode | |
+    | - | :-: | - | :-: | - | - |
+    | [top](#food--drink) | :crab: | \`:crab:\` | :shrimp: | \`:shrimp:\` | [top](#table-of-contents) |
+    | [top](#food--drink) | :squid: | \`:squid:\` | | | [top](#table-of-contents) |
+
+    #### Food Sweet
+
+    | | ico | shortcode | ico | shortcode | |
+    | - | :-: | - | :-: | - | - |
+    | [top](#food--drink) | :icecream: | \`:icecream:\` | :shaved_ice: | \`:shaved_ice:\` | [top](#table-of-contents) |
+    | [top](#food--drink) | :ice_cream: | \`:ice_cream:\` | :doughnut: | \`:doughnut:\` | [top](#table-of-contents) |
+    | [top](#food--drink) | :cookie: | \`:cookie:\` | :birthday: | \`:birthday:\` | [top](#table-of-contents) |
+    | [top](#food--drink) | :cake: | \`:cake:\` | :chocolate_bar: | \`:chocolate_bar:\` | [top](#table-of-contents) |
+    | [top](#food--drink) | :candy: | \`:candy:\` | :lollipop: | \`:lollipop:\` | [top](#table-of-contents) |
+    | [top](#food--drink) | :custard: | \`:custard:\` | :honey_pot: | \`:honey_pot:\` | [top](#table-of-contents) |
+
+    #### Drink
+
+    | | ico | shortcode | ico | shortcode | |
+    | - | :-: | - | :-: | - | - |
+    | [top](#food--drink) | :baby_bottle: | \`:baby_bottle:\` | :milk_glass: | \`:milk_glass:\` | [top](#table-of-contents) |
+    | [top](#food--drink) | :coffee: | \`:coffee:\` | :tea: | \`:tea:\` | [top](#table-of-contents) |
+    | [top](#food--drink) | :sake: | \`:sake:\` | :champagne: | \`:champagne:\` | [top](#table-of-contents) |
+    | [top](#food--drink) | :wine_glass: | \`:wine_glass:\` | :cocktail: | \`:cocktail:\` | [top](#table-of-contents) |
+    | [top](#food--drink) | :tropical_drink: | \`:tropical_drink:\` | :beer: | \`:beer:\` | [top](#table-of-contents) |
+    | [top](#food--drink) | :beers: | \`:beers:\` | :clinking_glasses: | \`:clinking_glasses:\` | [top](#table-of-contents) |
+    | [top](#food--drink) | :tumbler_glass: | \`:tumbler_glass:\` | | | [top](#table-of-contents) |
+
+    #### Dishware
+
+    | | ico | shortcode | ico | shortcode | |
+    | - | :-: | - | :-: | - | - |
+    | [top](#food--drink) | :plate_with_cutlery: | \`:plate_with_cutlery:\` | :fork_and_knife: | \`:fork_and_knife:\` | [top](#table-of-contents) |
+    | [top](#food--drink) | :spoon: | \`:spoon:\` | :hocho: | \`:hocho:\` <br /> \`:knife:\` | [top](#table-of-contents) |
+    | [top](#food--drink) | :amphora: | \`:amphora:\` | | | [top](#table-of-contents) |
+
+    ### Travel & Places
+
+    - [Place Map](#place-map)
+    - [Place Geographic](#place-geographic)
+    - [Place Building](#place-building)
+    - [Place Religious](#place-religious)
+    - [Place Other](#place-other)
+    - [Transport Ground](#transport-ground)
+    - [Transport Water](#transport-water)
+    - [Transport Air](#transport-air)
+    - [Hotel](#hotel)
+    - [Time](#time)
+    - [Sky & Weather](#sky--weather)
+
+    #### Place Map
+
+    | | ico | shortcode | ico | shortcode | |
+    | - | :-: | - | :-: | - | - |
+    | [top](#travel--places) | :earth_africa: | \`:earth_africa:\` | :earth_americas: | \`:earth_americas:\` | [top](#table-of-contents) |
+    | [top](#travel--places) | :earth_asia: | \`:earth_asia:\` | :globe_with_meridians: | \`:globe_with_meridians:\` | [top](#table-of-contents) |
+    | [top](#travel--places) | :world_map: | \`:world_map:\` | :japan: | \`:japan:\` | [top](#table-of-contents) |
+
+    #### Place Geographic
+
+    | | ico | shortcode | ico | shortcode | |
+    | - | :-: | - | :-: | - | - |
+    | [top](#travel--places) | :mountain_snow: | \`:mountain_snow:\` | :mountain: | \`:mountain:\` | [top](#table-of-contents) |
+    | [top](#travel--places) | :volcano: | \`:volcano:\` | :mount_fuji: | \`:mount_fuji:\` | [top](#table-of-contents) |
+    | [top](#travel--places) | :camping: | \`:camping:\` | :beach_umbrella: | \`:beach_umbrella:\` | [top](#table-of-contents) |
+    | [top](#travel--places) | :desert: | \`:desert:\` | :desert_island: | \`:desert_island:\` | [top](#table-of-contents) |
+    | [top](#travel--places) | :national_park: | \`:national_park:\` | | | [top](#table-of-contents) |
+
+    #### Place Building
+
+    | | ico | shortcode | ico | shortcode | |
+    | - | :-: | - | :-: | - | - |
+    | [top](#travel--places) | :stadium: | \`:stadium:\` | :classical_building: | \`:classical_building:\` | [top](#table-of-contents) |
+    | [top](#travel--places) | :building_construction: | \`:building_construction:\` | :houses: | \`:houses:\` | [top](#table-of-contents) |
+    | [top](#travel--places) | :derelict_house: | \`:derelict_house:\` | :house: | \`:house:\` | [top](#table-of-contents) |
+    | [top](#travel--places) | :house_with_garden: | \`:house_with_garden:\` | :office: | \`:office:\` | [top](#table-of-contents) |
+    | [top](#travel--places) | :post_office: | \`:post_office:\` | :european_post_office: | \`:european_post_office:\` | [top](#table-of-contents) |
+    | [top](#travel--places) | :hospital: | \`:hospital:\` | :bank: | \`:bank:\` | [top](#table-of-contents) |
+    | [top](#travel--places) | :hotel: | \`:hotel:\` | :love_hotel: | \`:love_hotel:\` | [top](#table-of-contents) |
+    | [top](#travel--places) | :convenience_store: | \`:convenience_store:\` | :school: | \`:school:\` | [top](#table-of-contents) |
+    | [top](#travel--places) | :department_store: | \`:department_store:\` | :factory: | \`:factory:\` | [top](#table-of-contents) |
+    | [top](#travel--places) | :japanese_castle: | \`:japanese_castle:\` | :european_castle: | \`:european_castle:\` | [top](#table-of-contents) |
+    | [top](#travel--places) | :wedding: | \`:wedding:\` | :tokyo_tower: | \`:tokyo_tower:\` | [top](#table-of-contents) |
+    | [top](#travel--places) | :statue_of_liberty: | \`:statue_of_liberty:\` | | | [top](#table-of-contents) |
+
+    #### Place Religious
+
+    | | ico | shortcode | ico | shortcode | |
+    | - | :-: | - | :-: | - | - |
+    | [top](#travel--places) | :church: | \`:church:\` | :mosque: | \`:mosque:\` | [top](#table-of-contents) |
+    | [top](#travel--places) | :synagogue: | \`:synagogue:\` | :shinto_shrine: | \`:shinto_shrine:\` | [top](#table-of-contents) |
+    | [top](#travel--places) | :kaaba: | \`:kaaba:\` | | | [top](#table-of-contents) |
+
+    #### Place Other
+
+    | | ico | shortcode | ico | shortcode | |
+    | - | :-: | - | :-: | - | - |
+    | [top](#travel--places) | :fountain: | \`:fountain:\` | :tent: | \`:tent:\` | [top](#table-of-contents) |
+    | [top](#travel--places) | :foggy: | \`:foggy:\` | :night_with_stars: | \`:night_with_stars:\` | [top](#table-of-contents) |
+    | [top](#travel--places) | :cityscape: | \`:cityscape:\` | :sunrise_over_mountains: | \`:sunrise_over_mountains:\` | [top](#table-of-contents) |
+    | [top](#travel--places) | :sunrise: | \`:sunrise:\` | :city_sunset: | \`:city_sunset:\` | [top](#table-of-contents) |
+    | [top](#travel--places) | :city_sunrise: | \`:city_sunrise:\` | :bridge_at_night: | \`:bridge_at_night:\` | [top](#table-of-contents) |
+    | [top](#travel--places) | :hotsprings: | \`:hotsprings:\` | :carousel_horse: | \`:carousel_horse:\` | [top](#table-of-contents) |
+    | [top](#travel--places) | :ferris_wheel: | \`:ferris_wheel:\` | :roller_coaster: | \`:roller_coaster:\` | [top](#table-of-contents) |
+    | [top](#travel--places) | :barber: | \`:barber:\` | :circus_tent: | \`:circus_tent:\` | [top](#table-of-contents) |
+
+    #### Transport Ground
+
+    | | ico | shortcode | ico | shortcode | |
+    | - | :-: | - | :-: | - | - |
+    | [top](#travel--places) | :steam_locomotive: | \`:steam_locomotive:\` | :railway_car: | \`:railway_car:\` | [top](#table-of-contents) |
+    | [top](#travel--places) | :bullettrain_side: | \`:bullettrain_side:\` | :bullettrain_front: | \`:bullettrain_front:\` | [top](#table-of-contents) |
+    | [top](#travel--places) | :train2: | \`:train2:\` | :metro: | \`:metro:\` | [top](#table-of-contents) |
+    | [top](#travel--places) | :light_rail: | \`:light_rail:\` | :station: | \`:station:\` | [top](#table-of-contents) |
+    | [top](#travel--places) | :tram: | \`:tram:\` | :monorail: | \`:monorail:\` | [top](#table-of-contents) |
+    | [top](#travel--places) | :mountain_railway: | \`:mountain_railway:\` | :train: | \`:train:\` | [top](#table-of-contents) |
+    | [top](#travel--places) | :bus: | \`:bus:\` | :oncoming_bus: | \`:oncoming_bus:\` | [top](#table-of-contents) |
+    | [top](#travel--places) | :trolleybus: | \`:trolleybus:\` | :minibus: | \`:minibus:\` | [top](#table-of-contents) |
+    | [top](#travel--places) | :ambulance: | \`:ambulance:\` | :fire_engine: | \`:fire_engine:\` | [top](#table-of-contents) |
+    | [top](#travel--places) | :police_car: | \`:police_car:\` | :oncoming_police_car: | \`:oncoming_police_car:\` | [top](#table-of-contents) |
+    | [top](#travel--places) | :taxi: | \`:taxi:\` | :oncoming_taxi: | \`:oncoming_taxi:\` | [top](#table-of-contents) |
+    | [top](#travel--places) | :car: | \`:car:\` <br /> \`:red_car:\` | :oncoming_automobile: | \`:oncoming_automobile:\` | [top](#table-of-contents) |
+    | [top](#travel--places) | :blue_car: | \`:blue_car:\` | :truck: | \`:truck:\` | [top](#table-of-contents) |
+    | [top](#travel--places) | :articulated_lorry: | \`:articulated_lorry:\` | :tractor: | \`:tractor:\` | [top](#table-of-contents) |
+    | [top](#travel--places) | :racing_car: | \`:racing_car:\` | :motorcycle: | \`:motorcycle:\` | [top](#table-of-contents) |
+    | [top](#travel--places) | :motor_scooter: | \`:motor_scooter:\` | :bike: | \`:bike:\` | [top](#table-of-contents) |
+    | [top](#travel--places) | :kick_scooter: | \`:kick_scooter:\` | :busstop: | \`:busstop:\` | [top](#table-of-contents) |
+    | [top](#travel--places) | :motorway: | \`:motorway:\` | :railway_track: | \`:railway_track:\` | [top](#table-of-contents) |
+    | [top](#travel--places) | :oil_drum: | \`:oil_drum:\` | :fuelpump: | \`:fuelpump:\` | [top](#table-of-contents) |
+    | [top](#travel--places) | :rotating_light: | \`:rotating_light:\` | :traffic_light: | \`:traffic_light:\` | [top](#table-of-contents) |
+    | [top](#travel--places) | :vertical_traffic_light: | \`:vertical_traffic_light:\` | :stop_sign: | \`:stop_sign:\` | [top](#table-of-contents) |
+    | [top](#travel--places) | :construction: | \`:construction:\` | | | [top](#table-of-contents) |
+
+    #### Transport Water
+
+    | | ico | shortcode | ico | shortcode | |
+    | - | :-: | - | :-: | - | - |
+    | [top](#travel--places) | :anchor: | \`:anchor:\` | :boat: | \`:boat:\` <br /> \`:sailboat:\` | [top](#table-of-contents) |
+    | [top](#travel--places) | :canoe: | \`:canoe:\` | :speedboat: | \`:speedboat:\` | [top](#table-of-contents) |
+    | [top](#travel--places) | :passenger_ship: | \`:passenger_ship:\` | :ferry: | \`:ferry:\` | [top](#table-of-contents) |
+    | [top](#travel--places) | :motor_boat: | \`:motor_boat:\` | :ship: | \`:ship:\` | [top](#table-of-contents) |
+
+    #### Transport Air
+
+    | | ico | shortcode | ico | shortcode | |
+    | - | :-: | - | :-: | - | - |
+    | [top](#travel--places) | :airplane: | \`:airplane:\` | :small_airplane: | \`:small_airplane:\` | [top](#table-of-contents) |
+    | [top](#travel--places) | :flight_departure: | \`:flight_departure:\` | :flight_arrival: | \`:flight_arrival:\` | [top](#table-of-contents) |
+    | [top](#travel--places) | :seat: | \`:seat:\` | :helicopter: | \`:helicopter:\` | [top](#table-of-contents) |
+    | [top](#travel--places) | :suspension_railway: | \`:suspension_railway:\` | :mountain_cableway: | \`:mountain_cableway:\` | [top](#table-of-contents) |
+    | [top](#travel--places) | :aerial_tramway: | \`:aerial_tramway:\` | :artificial_satellite: | \`:artificial_satellite:\` | [top](#table-of-contents) |
+    | [top](#travel--places) | :rocket: | \`:rocket:\` | | | [top](#table-of-contents) |
+
+    #### Hotel
+
+    | | ico | shortcode | |
+    | - | :-: | - | - |
+    | [top](#travel--places) | :bellhop_bell: | \`:bellhop_bell:\` | [top](#table-of-contents) |
+
+    #### Time
+
+    | | ico | shortcode | ico | shortcode | |
+    | - | :-: | - | :-: | - | - |
+    | [top](#travel--places) | :hourglass: | \`:hourglass:\` | :hourglass_flowing_sand: | \`:hourglass_flowing_sand:\` | [top](#table-of-contents) |
+    | [top](#travel--places) | :watch: | \`:watch:\` | :alarm_clock: | \`:alarm_clock:\` | [top](#table-of-contents) |
+    | [top](#travel--places) | :stopwatch: | \`:stopwatch:\` | :timer_clock: | \`:timer_clock:\` | [top](#table-of-contents) |
+    | [top](#travel--places) | :mantelpiece_clock: | \`:mantelpiece_clock:\` | :clock12: | \`:clock12:\` | [top](#table-of-contents) |
+    | [top](#travel--places) | :clock1230: | \`:clock1230:\` | :clock1: | \`:clock1:\` | [top](#table-of-contents) |
+    | [top](#travel--places) | :clock130: | \`:clock130:\` | :clock2: | \`:clock2:\` | [top](#table-of-contents) |
+    | [top](#travel--places) | :clock230: | \`:clock230:\` | :clock3: | \`:clock3:\` | [top](#table-of-contents) |
+    | [top](#travel--places) | :clock330: | \`:clock330:\` | :clock4: | \`:clock4:\` | [top](#table-of-contents) |
+    | [top](#travel--places) | :clock430: | \`:clock430:\` | :clock5: | \`:clock5:\` | [top](#table-of-contents) |
+    | [top](#travel--places) | :clock530: | \`:clock530:\` | :clock6: | \`:clock6:\` | [top](#table-of-contents) |
+    | [top](#travel--places) | :clock630: | \`:clock630:\` | :clock7: | \`:clock7:\` | [top](#table-of-contents) |
+    | [top](#travel--places) | :clock730: | \`:clock730:\` | :clock8: | \`:clock8:\` | [top](#table-of-contents) |
+    | [top](#travel--places) | :clock830: | \`:clock830:\` | :clock9: | \`:clock9:\` | [top](#table-of-contents) |
+    | [top](#travel--places) | :clock930: | \`:clock930:\` | :clock10: | \`:clock10:\` | [top](#table-of-contents) |
+    | [top](#travel--places) | :clock1030: | \`:clock1030:\` | :clock11: | \`:clock11:\` | [top](#table-of-contents) |
+    | [top](#travel--places) | :clock1130: | \`:clock1130:\` | | | [top](#table-of-contents) |
+
+    #### Sky & Weather
+
+    | | ico | shortcode | ico | shortcode | |
+    | - | :-: | - | :-: | - | - |
+    | [top](#travel--places) | :new_moon: | \`:new_moon:\` | :waxing_crescent_moon: | \`:waxing_crescent_moon:\` | [top](#table-of-contents) |
+    | [top](#travel--places) | :first_quarter_moon: | \`:first_quarter_moon:\` | :moon: | \`:moon:\` <br /> \`:waxing_gibbous_moon:\` | [top](#table-of-contents) |
+    | [top](#travel--places) | :full_moon: | \`:full_moon:\` | :waning_gibbous_moon: | \`:waning_gibbous_moon:\` | [top](#table-of-contents) |
+    | [top](#travel--places) | :last_quarter_moon: | \`:last_quarter_moon:\` | :waning_crescent_moon: | \`:waning_crescent_moon:\` | [top](#table-of-contents) |
+    | [top](#travel--places) | :crescent_moon: | \`:crescent_moon:\` | :new_moon_with_face: | \`:new_moon_with_face:\` | [top](#table-of-contents) |
+    | [top](#travel--places) | :first_quarter_moon_with_face: | \`:first_quarter_moon_with_face:\` | :last_quarter_moon_with_face: | \`:last_quarter_moon_with_face:\` | [top](#table-of-contents) |
+    | [top](#travel--places) | :thermometer: | \`:thermometer:\` | :sunny: | \`:sunny:\` | [top](#table-of-contents) |
+    | [top](#travel--places) | :full_moon_with_face: | \`:full_moon_with_face:\` | :sun_with_face: | \`:sun_with_face:\` | [top](#table-of-contents) |
+    | [top](#travel--places) | :star: | \`:star:\` | :star2: | \`:star2:\` | [top](#table-of-contents) |
+    | [top](#travel--places) | :stars: | \`:stars:\` | :milky_way: | \`:milky_way:\` | [top](#table-of-contents) |
+    | [top](#travel--places) | :cloud: | \`:cloud:\` | :partly_sunny: | \`:partly_sunny:\` | [top](#table-of-contents) |
+    | [top](#travel--places) | :cloud_with_lightning_and_rain: | \`:cloud_with_lightning_and_rain:\` | :sun_behind_small_cloud: | \`:sun_behind_small_cloud:\` | [top](#table-of-contents) |
+    | [top](#travel--places) | :sun_behind_large_cloud: | \`:sun_behind_large_cloud:\` | :sun_behind_rain_cloud: | \`:sun_behind_rain_cloud:\` | [top](#table-of-contents) |
+    | [top](#travel--places) | :cloud_with_rain: | \`:cloud_with_rain:\` | :cloud_with_snow: | \`:cloud_with_snow:\` | [top](#table-of-contents) |
+    | [top](#travel--places) | :cloud_with_lightning: | \`:cloud_with_lightning:\` | :tornado: | \`:tornado:\` | [top](#table-of-contents) |
+    | [top](#travel--places) | :fog: | \`:fog:\` | :wind_face: | \`:wind_face:\` | [top](#table-of-contents) |
+    | [top](#travel--places) | :cyclone: | \`:cyclone:\` | :rainbow: | \`:rainbow:\` | [top](#table-of-contents) |
+    | [top](#travel--places) | :closed_umbrella: | \`:closed_umbrella:\` | :open_umbrella: | \`:open_umbrella:\` | [top](#table-of-contents) |
+    | [top](#travel--places) | :umbrella: | \`:umbrella:\` | :parasol_on_ground: | \`:parasol_on_ground:\` | [top](#table-of-contents) |
+    | [top](#travel--places) | :zap: | \`:zap:\` | :snowflake: | \`:snowflake:\` | [top](#table-of-contents) |
+    | [top](#travel--places) | :snowman_with_snow: | \`:snowman_with_snow:\` | :snowman: | \`:snowman:\` | [top](#table-of-contents) |
+    | [top](#travel--places) | :comet: | \`:comet:\` | :fire: | \`:fire:\` | [top](#table-of-contents) |
+    | [top](#travel--places) | :droplet: | \`:droplet:\` | :ocean: | \`:ocean:\` | [top](#table-of-contents) |
+
+    ### Activities
+
+    - [Event](#event)
+    - [Award Medal](#award-medal)
+    - [Sport](#sport)
+    - [Game](#game)
+    - [Arts & Crafts](#arts--crafts)
+
+    #### Event
+
+    | | ico | shortcode | ico | shortcode | |
+    | - | :-: | - | :-: | - | - |
+    | [top](#activities) | :jack_o_lantern: | \`:jack_o_lantern:\` | :christmas_tree: | \`:christmas_tree:\` | [top](#table-of-contents) |
+    | [top](#activities) | :fireworks: | \`:fireworks:\` | :sparkler: | \`:sparkler:\` | [top](#table-of-contents) |
+    | [top](#activities) | :sparkles: | \`:sparkles:\` | :balloon: | \`:balloon:\` | [top](#table-of-contents) |
+    | [top](#activities) | :tada: | \`:tada:\` | :confetti_ball: | \`:confetti_ball:\` | [top](#table-of-contents) |
+    | [top](#activities) | :tanabata_tree: | \`:tanabata_tree:\` | :bamboo: | \`:bamboo:\` | [top](#table-of-contents) |
+    | [top](#activities) | :dolls: | \`:dolls:\` | :flags: | \`:flags:\` | [top](#table-of-contents) |
+    | [top](#activities) | :wind_chime: | \`:wind_chime:\` | :rice_scene: | \`:rice_scene:\` | [top](#table-of-contents) |
+    | [top](#activities) | :ribbon: | \`:ribbon:\` | :gift: | \`:gift:\` | [top](#table-of-contents) |
+    | [top](#activities) | :reminder_ribbon: | \`:reminder_ribbon:\` | :tickets: | \`:tickets:\` | [top](#table-of-contents) |
+    | [top](#activities) | :ticket: | \`:ticket:\` | | | [top](#table-of-contents) |
+
+    #### Award Medal
+
+    | | ico | shortcode | ico | shortcode | |
+    | - | :-: | - | :-: | - | - |
+    | [top](#activities) | :medal_military: | \`:medal_military:\` | :trophy: | \`:trophy:\` | [top](#table-of-contents) |
+    | [top](#activities) | :medal_sports: | \`:medal_sports:\` | :1st_place_medal: | \`:1st_place_medal:\` | [top](#table-of-contents) |
+    | [top](#activities) | :2nd_place_medal: | \`:2nd_place_medal:\` | :3rd_place_medal: | \`:3rd_place_medal:\` | [top](#table-of-contents) |
+
+    #### Sport
+
+    | | ico | shortcode | ico | shortcode | |
+    | - | :-: | - | :-: | - | - |
+    | [top](#activities) | :soccer: | \`:soccer:\` | :baseball: | \`:baseball:\` | [top](#table-of-contents) |
+    | [top](#activities) | :basketball: | \`:basketball:\` | :volleyball: | \`:volleyball:\` | [top](#table-of-contents) |
+    | [top](#activities) | :football: | \`:football:\` | :rugby_football: | \`:rugby_football:\` | [top](#table-of-contents) |
+    | [top](#activities) | :tennis: | \`:tennis:\` | :bowling: | \`:bowling:\` | [top](#table-of-contents) |
+    | [top](#activities) | :cricket: | \`:cricket:\` | :field_hockey: | \`:field_hockey:\` | [top](#table-of-contents) |
+    | [top](#activities) | :ice_hockey: | \`:ice_hockey:\` | :ping_pong: | \`:ping_pong:\` | [top](#table-of-contents) |
+    | [top](#activities) | :badminton: | \`:badminton:\` | :boxing_glove: | \`:boxing_glove:\` | [top](#table-of-contents) |
+    | [top](#activities) | :martial_arts_uniform: | \`:martial_arts_uniform:\` | :goal_net: | \`:goal_net:\` | [top](#table-of-contents) |
+    | [top](#activities) | :golf: | \`:golf:\` | :ice_skate: | \`:ice_skate:\` | [top](#table-of-contents) |
+    | [top](#activities) | :fishing_pole_and_fish: | \`:fishing_pole_and_fish:\` | :running_shirt_with_sash: | \`:running_shirt_with_sash:\` | [top](#table-of-contents) |
+    | [top](#activities) | :ski: | \`:ski:\` | | | [top](#table-of-contents) |
+
+    #### Game
+
+    | | ico | shortcode | ico | shortcode | |
+    | - | :-: | - | :-: | - | - |
+    | [top](#activities) | :dart: | \`:dart:\` | :8ball: | \`:8ball:\` | [top](#table-of-contents) |
+    | [top](#activities) | :crystal_ball: | \`:crystal_ball:\` | :video_game: | \`:video_game:\` | [top](#table-of-contents) |
+    | [top](#activities) | :joystick: | \`:joystick:\` | :slot_machine: | \`:slot_machine:\` | [top](#table-of-contents) |
+    | [top](#activities) | :game_die: | \`:game_die:\` | :spades: | \`:spades:\` | [top](#table-of-contents) |
+    | [top](#activities) | :hearts: | \`:hearts:\` | :diamonds: | \`:diamonds:\` | [top](#table-of-contents) |
+    | [top](#activities) | :clubs: | \`:clubs:\` | :black_joker: | \`:black_joker:\` | [top](#table-of-contents) |
+    | [top](#activities) | :mahjong: | \`:mahjong:\` | :flower_playing_cards: | \`:flower_playing_cards:\` | [top](#table-of-contents) |
+
+    #### Arts & Crafts
+
+    | | ico | shortcode | ico | shortcode | |
+    | - | :-: | - | :-: | - | - |
+    | [top](#activities) | :performing_arts: | \`:performing_arts:\` | :framed_picture: | \`:framed_picture:\` | [top](#table-of-contents) |
+    | [top](#activities) | :art: | \`:art:\` | | | [top](#table-of-contents) |
+
+    ### Objects
+
+    - [Clothing](#clothing)
+    - [Sound](#sound)
+    - [Music](#music)
+    - [Musical Instrument](#musical-instrument)
+    - [Phone](#phone)
+    - [Computer](#computer)
+    - [Light & Video](#light--video)
+    - [Book Paper](#book-paper)
+    - [Money](#money)
+    - [Mail](#mail)
+    - [Writing](#writing)
+    - [Office](#office)
+    - [Lock](#lock)
+    - [Tool](#tool)
+    - [Science](#science)
+    - [Medical](#medical)
+    - [Household](#household)
+    - [Other Object](#other-object)
+
+    #### Clothing
+
+    | | ico | shortcode | ico | shortcode | |
+    | - | :-: | - | :-: | - | - |
+    | [top](#objects) | :eyeglasses: | \`:eyeglasses:\` | :dark_sunglasses: | \`:dark_sunglasses:\` | [top](#table-of-contents) |
+    | [top](#objects) | :necktie: | \`:necktie:\` | :shirt: | \`:shirt:\` <br /> \`:tshirt:\` | [top](#table-of-contents) |
+    | [top](#objects) | :jeans: | \`:jeans:\` | :dress: | \`:dress:\` | [top](#table-of-contents) |
+    | [top](#objects) | :kimono: | \`:kimono:\` | :bikini: | \`:bikini:\` | [top](#table-of-contents) |
+    | [top](#objects) | :womans_clothes: | \`:womans_clothes:\` | :purse: | \`:purse:\` | [top](#table-of-contents) |
+    | [top](#objects) | :handbag: | \`:handbag:\` | :pouch: | \`:pouch:\` | [top](#table-of-contents) |
+    | [top](#objects) | :shopping: | \`:shopping:\` | :school_satchel: | \`:school_satchel:\` | [top](#table-of-contents) |
+    | [top](#objects) | :mans_shoe: | \`:mans_shoe:\` <br /> \`:shoe:\` | :athletic_shoe: | \`:athletic_shoe:\` | [top](#table-of-contents) |
+    | [top](#objects) | :high_heel: | \`:high_heel:\` | :sandal: | \`:sandal:\` | [top](#table-of-contents) |
+    | [top](#objects) | :boot: | \`:boot:\` | :crown: | \`:crown:\` | [top](#table-of-contents) |
+    | [top](#objects) | :womans_hat: | \`:womans_hat:\` | :tophat: | \`:tophat:\` | [top](#table-of-contents) |
+    | [top](#objects) | :mortar_board: | \`:mortar_board:\` | :rescue_worker_helmet: | \`:rescue_worker_helmet:\` | [top](#table-of-contents) |
+    | [top](#objects) | :prayer_beads: | \`:prayer_beads:\` | :lipstick: | \`:lipstick:\` | [top](#table-of-contents) |
+    | [top](#objects) | :ring: | \`:ring:\` | :gem: | \`:gem:\` | [top](#table-of-contents) |
+
+    #### Sound
+
+    | | ico | shortcode | ico | shortcode | |
+    | - | :-: | - | :-: | - | - |
+    | [top](#objects) | :mute: | \`:mute:\` | :speaker: | \`:speaker:\` | [top](#table-of-contents) |
+    | [top](#objects) | :sound: | \`:sound:\` | :loud_sound: | \`:loud_sound:\` | [top](#table-of-contents) |
+    | [top](#objects) | :loudspeaker: | \`:loudspeaker:\` | :mega: | \`:mega:\` | [top](#table-of-contents) |
+    | [top](#objects) | :postal_horn: | \`:postal_horn:\` | :bell: | \`:bell:\` | [top](#table-of-contents) |
+    | [top](#objects) | :no_bell: | \`:no_bell:\` | | | [top](#table-of-contents) |
+
+    #### Music
+
+    | | ico | shortcode | ico | shortcode | |
+    | - | :-: | - | :-: | - | - |
+    | [top](#objects) | :musical_score: | \`:musical_score:\` | :musical_note: | \`:musical_note:\` | [top](#table-of-contents) |
+    | [top](#objects) | :notes: | \`:notes:\` | :studio_microphone: | \`:studio_microphone:\` | [top](#table-of-contents) |
+    | [top](#objects) | :level_slider: | \`:level_slider:\` | :control_knobs: | \`:control_knobs:\` | [top](#table-of-contents) |
+    | [top](#objects) | :microphone: | \`:microphone:\` | :headphones: | \`:headphones:\` | [top](#table-of-contents) |
+    | [top](#objects) | :radio: | \`:radio:\` | | | [top](#table-of-contents) |
+
+    #### Musical Instrument
+
+    | | ico | shortcode | ico | shortcode | |
+    | - | :-: | - | :-: | - | - |
+    | [top](#objects) | :saxophone: | \`:saxophone:\` | :guitar: | \`:guitar:\` | [top](#table-of-contents) |
+    | [top](#objects) | :musical_keyboard: | \`:musical_keyboard:\` | :trumpet: | \`:trumpet:\` | [top](#table-of-contents) |
+    | [top](#objects) | :violin: | \`:violin:\` | :drum: | \`:drum:\` | [top](#table-of-contents) |
+
+    #### Phone
+
+    | | ico | shortcode | ico | shortcode | |
+    | - | :-: | - | :-: | - | - |
+    | [top](#objects) | :iphone: | \`:iphone:\` | :calling: | \`:calling:\` | [top](#table-of-contents) |
+    | [top](#objects) | :phone: | \`:phone:\` <br /> \`:telephone:\` | :telephone_receiver: | \`:telephone_receiver:\` | [top](#table-of-contents) |
+    | [top](#objects) | :pager: | \`:pager:\` | :fax: | \`:fax:\` | [top](#table-of-contents) |
+
+    #### Computer
+
+    | | ico | shortcode | ico | shortcode | |
+    | - | :-: | - | :-: | - | - |
+    | [top](#objects) | :battery: | \`:battery:\` | :electric_plug: | \`:electric_plug:\` | [top](#table-of-contents) |
+    | [top](#objects) | :computer: | \`:computer:\` | :desktop_computer: | \`:desktop_computer:\` | [top](#table-of-contents) |
+    | [top](#objects) | :printer: | \`:printer:\` | :keyboard: | \`:keyboard:\` | [top](#table-of-contents) |
+    | [top](#objects) | :computer_mouse: | \`:computer_mouse:\` | :trackball: | \`:trackball:\` | [top](#table-of-contents) |
+    | [top](#objects) | :minidisc: | \`:minidisc:\` | :floppy_disk: | \`:floppy_disk:\` | [top](#table-of-contents) |
+    | [top](#objects) | :cd: | \`:cd:\` | :dvd: | \`:dvd:\` | [top](#table-of-contents) |
+
+    #### Light & Video
+
+    | | ico | shortcode | ico | shortcode | |
+    | - | :-: | - | :-: | - | - |
+    | [top](#objects) | :movie_camera: | \`:movie_camera:\` | :film_strip: | \`:film_strip:\` | [top](#table-of-contents) |
+    | [top](#objects) | :film_projector: | \`:film_projector:\` | :clapper: | \`:clapper:\` | [top](#table-of-contents) |
+    | [top](#objects) | :tv: | \`:tv:\` | :camera: | \`:camera:\` | [top](#table-of-contents) |
+    | [top](#objects) | :camera_flash: | \`:camera_flash:\` | :video_camera: | \`:video_camera:\` | [top](#table-of-contents) |
+    | [top](#objects) | :vhs: | \`:vhs:\` | :mag: | \`:mag:\` | [top](#table-of-contents) |
+    | [top](#objects) | :mag_right: | \`:mag_right:\` | :candle: | \`:candle:\` | [top](#table-of-contents) |
+    | [top](#objects) | :bulb: | \`:bulb:\` | :flashlight: | \`:flashlight:\` | [top](#table-of-contents) |
+    | [top](#objects) | :izakaya_lantern: | \`:izakaya_lantern:\` <br /> \`:lantern:\` | | | [top](#table-of-contents) |
+
+    #### Book Paper
+
+    | | ico | shortcode | ico | shortcode | |
+    | - | :-: | - | :-: | - | - |
+    | [top](#objects) | :notebook_with_decorative_cover: | \`:notebook_with_decorative_cover:\` | :closed_book: | \`:closed_book:\` | [top](#table-of-contents) |
+    | [top](#objects) | :book: | \`:book:\` <br /> \`:open_book:\` | :green_book: | \`:green_book:\` | [top](#table-of-contents) |
+    | [top](#objects) | :blue_book: | \`:blue_book:\` | :orange_book: | \`:orange_book:\` | [top](#table-of-contents) |
+    | [top](#objects) | :books: | \`:books:\` | :notebook: | \`:notebook:\` | [top](#table-of-contents) |
+    | [top](#objects) | :ledger: | \`:ledger:\` | :page_with_curl: | \`:page_with_curl:\` | [top](#table-of-contents) |
+    | [top](#objects) | :scroll: | \`:scroll:\` | :page_facing_up: | \`:page_facing_up:\` | [top](#table-of-contents) |
+    | [top](#objects) | :newspaper: | \`:newspaper:\` | :newspaper_roll: | \`:newspaper_roll:\` | [top](#table-of-contents) |
+    | [top](#objects) | :bookmark_tabs: | \`:bookmark_tabs:\` | :bookmark: | \`:bookmark:\` | [top](#table-of-contents) |
+    | [top](#objects) | :label: | \`:label:\` | | | [top](#table-of-contents) |
+
+    #### Money
+
+    | | ico | shortcode | ico | shortcode | |
+    | - | :-: | - | :-: | - | - |
+    | [top](#objects) | :moneybag: | \`:moneybag:\` | :yen: | \`:yen:\` | [top](#table-of-contents) |
+    | [top](#objects) | :dollar: | \`:dollar:\` | :euro: | \`:euro:\` | [top](#table-of-contents) |
+    | [top](#objects) | :pound: | \`:pound:\` | :money_with_wings: | \`:money_with_wings:\` | [top](#table-of-contents) |
+    | [top](#objects) | :credit_card: | \`:credit_card:\` | :chart: | \`:chart:\` | [top](#table-of-contents) |
+    | [top](#objects) | :currency_exchange: | \`:currency_exchange:\` | :heavy_dollar_sign: | \`:heavy_dollar_sign:\` | [top](#table-of-contents) |
+
+    #### Mail
+
+    | | ico | shortcode | ico | shortcode | |
+    | - | :-: | - | :-: | - | - |
+    | [top](#objects) | :email: | \`:email:\` <br /> \`:envelope:\` | :e-mail: | \`:e-mail:\` | [top](#table-of-contents) |
+    | [top](#objects) | :incoming_envelope: | \`:incoming_envelope:\` | :envelope_with_arrow: | \`:envelope_with_arrow:\` | [top](#table-of-contents) |
+    | [top](#objects) | :outbox_tray: | \`:outbox_tray:\` | :inbox_tray: | \`:inbox_tray:\` | [top](#table-of-contents) |
+    | [top](#objects) | :package: | \`:package:\` | :mailbox: | \`:mailbox:\` | [top](#table-of-contents) |
+    | [top](#objects) | :mailbox_closed: | \`:mailbox_closed:\` | :mailbox_with_mail: | \`:mailbox_with_mail:\` | [top](#table-of-contents) |
+    | [top](#objects) | :mailbox_with_no_mail: | \`:mailbox_with_no_mail:\` | :postbox: | \`:postbox:\` | [top](#table-of-contents) |
+    | [top](#objects) | :ballot_box: | \`:ballot_box:\` | | | [top](#table-of-contents) |
+
+    #### Writing
+
+    | | ico | shortcode | ico | shortcode | |
+    | - | :-: | - | :-: | - | - |
+    | [top](#objects) | :pencil2: | \`:pencil2:\` | :black_nib: | \`:black_nib:\` | [top](#table-of-contents) |
+    | [top](#objects) | :fountain_pen: | \`:fountain_pen:\` | :pen: | \`:pen:\` | [top](#table-of-contents) |
+    | [top](#objects) | :paintbrush: | \`:paintbrush:\` | :crayon: | \`:crayon:\` | [top](#table-of-contents) |
+    | [top](#objects) | :memo: | \`:memo:\` <br /> \`:pencil:\` | | | [top](#table-of-contents) |
+
+    #### Office
+
+    | | ico | shortcode | ico | shortcode | |
+    | - | :-: | - | :-: | - | - |
+    | [top](#objects) | :briefcase: | \`:briefcase:\` | :file_folder: | \`:file_folder:\` | [top](#table-of-contents) |
+    | [top](#objects) | :open_file_folder: | \`:open_file_folder:\` | :card_index_dividers: | \`:card_index_dividers:\` | [top](#table-of-contents) |
+    | [top](#objects) | :date: | \`:date:\` | :calendar: | \`:calendar:\` | [top](#table-of-contents) |
+    | [top](#objects) | :spiral_notepad: | \`:spiral_notepad:\` | :spiral_calendar: | \`:spiral_calendar:\` | [top](#table-of-contents) |
+    | [top](#objects) | :card_index: | \`:card_index:\` | :chart_with_upwards_trend: | \`:chart_with_upwards_trend:\` | [top](#table-of-contents) |
+    | [top](#objects) | :chart_with_downwards_trend: | \`:chart_with_downwards_trend:\` | :bar_chart: | \`:bar_chart:\` | [top](#table-of-contents) |
+    | [top](#objects) | :clipboard: | \`:clipboard:\` | :pushpin: | \`:pushpin:\` | [top](#table-of-contents) |
+    | [top](#objects) | :round_pushpin: | \`:round_pushpin:\` | :paperclip: | \`:paperclip:\` | [top](#table-of-contents) |
+    | [top](#objects) | :paperclips: | \`:paperclips:\` | :straight_ruler: | \`:straight_ruler:\` | [top](#table-of-contents) |
+    | [top](#objects) | :triangular_ruler: | \`:triangular_ruler:\` | :scissors: | \`:scissors:\` | [top](#table-of-contents) |
+    | [top](#objects) | :card_file_box: | \`:card_file_box:\` | :file_cabinet: | \`:file_cabinet:\` | [top](#table-of-contents) |
+    | [top](#objects) | :wastebasket: | \`:wastebasket:\` | | | [top](#table-of-contents) |
+
+    #### Lock
+
+    | | ico | shortcode | ico | shortcode | |
+    | - | :-: | - | :-: | - | - |
+    | [top](#objects) | :lock: | \`:lock:\` | :unlock: | \`:unlock:\` | [top](#table-of-contents) |
+    | [top](#objects) | :lock_with_ink_pen: | \`:lock_with_ink_pen:\` | :closed_lock_with_key: | \`:closed_lock_with_key:\` | [top](#table-of-contents) |
+    | [top](#objects) | :key: | \`:key:\` | :old_key: | \`:old_key:\` | [top](#table-of-contents) |
+
+    #### Tool
+
+    | | ico | shortcode | ico | shortcode | |
+    | - | :-: | - | :-: | - | - |
+    | [top](#objects) | :hammer: | \`:hammer:\` | :pick: | \`:pick:\` | [top](#table-of-contents) |
+    | [top](#objects) | :hammer_and_pick: | \`:hammer_and_pick:\` | :hammer_and_wrench: | \`:hammer_and_wrench:\` | [top](#table-of-contents) |
+    | [top](#objects) | :dagger: | \`:dagger:\` | :crossed_swords: | \`:crossed_swords:\` | [top](#table-of-contents) |
+    | [top](#objects) | :gun: | \`:gun:\` | :bow_and_arrow: | \`:bow_and_arrow:\` | [top](#table-of-contents) |
+    | [top](#objects) | :shield: | \`:shield:\` | :wrench: | \`:wrench:\` | [top](#table-of-contents) |
+    | [top](#objects) | :nut_and_bolt: | \`:nut_and_bolt:\` | :gear: | \`:gear:\` | [top](#table-of-contents) |
+    | [top](#objects) | :clamp: | \`:clamp:\` | :balance_scale: | \`:balance_scale:\` | [top](#table-of-contents) |
+    | [top](#objects) | :link: | \`:link:\` | :chains: | \`:chains:\` | [top](#table-of-contents) |
+
+    #### Science
+
+    | | ico | shortcode | ico | shortcode | |
+    | - | :-: | - | :-: | - | - |
+    | [top](#objects) | :alembic: | \`:alembic:\` | :microscope: | \`:microscope:\` | [top](#table-of-contents) |
+    | [top](#objects) | :telescope: | \`:telescope:\` | :satellite: | \`:satellite:\` | [top](#table-of-contents) |
+
+    #### Medical
+
+    | | ico | shortcode | ico | shortcode | |
+    | - | :-: | - | :-: | - | - |
+    | [top](#objects) | :syringe: | \`:syringe:\` | :pill: | \`:pill:\` | [top](#table-of-contents) |
+
+    #### Household
+
+    | | ico | shortcode | ico | shortcode | |
+    | - | :-: | - | :-: | - | - |
+    | [top](#objects) | :door: | \`:door:\` | :bed: | \`:bed:\` | [top](#table-of-contents) |
+    | [top](#objects) | :couch_and_lamp: | \`:couch_and_lamp:\` | :toilet: | \`:toilet:\` | [top](#table-of-contents) |
+    | [top](#objects) | :shower: | \`:shower:\` | :bathtub: | \`:bathtub:\` | [top](#table-of-contents) |
+    | [top](#objects) | :shopping_cart: | \`:shopping_cart:\` | | | [top](#table-of-contents) |
+
+    #### Other Object
+
+    | | ico | shortcode | ico | shortcode | |
+    | - | :-: | - | :-: | - | - |
+    | [top](#objects) | :smoking: | \`:smoking:\` | :coffin: | \`:coffin:\` | [top](#table-of-contents) |
+    | [top](#objects) | :funeral_urn: | \`:funeral_urn:\` | :moyai: | \`:moyai:\` | [top](#table-of-contents) |
+
+    ### Symbols
+
+    - [Transport Sign](#transport-sign)
+    - [Warning](#warning)
+    - [Arrow](#arrow)
+    - [Religion](#religion)
+    - [Zodiac](#zodiac)
+    - [Av Symbol](#av-symbol)
+    - [Other Symbol](#other-symbol)
+    - [Keycap](#keycap)
+    - [Alphanum](#alphanum)
+    - [Geometric](#geometric)
+
+    #### Transport Sign
+
+    | | ico | shortcode | ico | shortcode | |
+    | - | :-: | - | :-: | - | - |
+    | [top](#symbols) | :atm: | \`:atm:\` | :put_litter_in_its_place: | \`:put_litter_in_its_place:\` | [top](#table-of-contents) |
+    | [top](#symbols) | :potable_water: | \`:potable_water:\` | :wheelchair: | \`:wheelchair:\` | [top](#table-of-contents) |
+    | [top](#symbols) | :mens: | \`:mens:\` | :womens: | \`:womens:\` | [top](#table-of-contents) |
+    | [top](#symbols) | :restroom: | \`:restroom:\` | :baby_symbol: | \`:baby_symbol:\` | [top](#table-of-contents) |
+    | [top](#symbols) | :wc: | \`:wc:\` | :passport_control: | \`:passport_control:\` | [top](#table-of-contents) |
+    | [top](#symbols) | :customs: | \`:customs:\` | :baggage_claim: | \`:baggage_claim:\` | [top](#table-of-contents) |
+    | [top](#symbols) | :left_luggage: | \`:left_luggage:\` | | | [top](#table-of-contents) |
+
+    #### Warning
+
+    | | ico | shortcode | ico | shortcode | |
+    | - | :-: | - | :-: | - | - |
+    | [top](#symbols) | :warning: | \`:warning:\` | :children_crossing: | \`:children_crossing:\` | [top](#table-of-contents) |
+    | [top](#symbols) | :no_entry: | \`:no_entry:\` | :no_entry_sign: | \`:no_entry_sign:\` | [top](#table-of-contents) |
+    | [top](#symbols) | :no_bicycles: | \`:no_bicycles:\` | :no_smoking: | \`:no_smoking:\` | [top](#table-of-contents) |
+    | [top](#symbols) | :do_not_litter: | \`:do_not_litter:\` | :non-potable_water: | \`:non-potable_water:\` | [top](#table-of-contents) |
+    | [top](#symbols) | :no_pedestrians: | \`:no_pedestrians:\` | :no_mobile_phones: | \`:no_mobile_phones:\` | [top](#table-of-contents) |
+    | [top](#symbols) | :underage: | \`:underage:\` | :radioactive: | \`:radioactive:\` | [top](#table-of-contents) |
+    | [top](#symbols) | :biohazard: | \`:biohazard:\` | | | [top](#table-of-contents) |
+
+    #### Arrow
+
+    | | ico | shortcode | ico | shortcode | |
+    | - | :-: | - | :-: | - | - |
+    | [top](#symbols) | :arrow_up: | \`:arrow_up:\` | :arrow_upper_right: | \`:arrow_upper_right:\` | [top](#table-of-contents) |
+    | [top](#symbols) | :arrow_right: | \`:arrow_right:\` | :arrow_lower_right: | \`:arrow_lower_right:\` | [top](#table-of-contents) |
+    | [top](#symbols) | :arrow_down: | \`:arrow_down:\` | :arrow_lower_left: | \`:arrow_lower_left:\` | [top](#table-of-contents) |
+    | [top](#symbols) | :arrow_left: | \`:arrow_left:\` | :arrow_upper_left: | \`:arrow_upper_left:\` | [top](#table-of-contents) |
+    | [top](#symbols) | :arrow_up_down: | \`:arrow_up_down:\` | :left_right_arrow: | \`:left_right_arrow:\` | [top](#table-of-contents) |
+    | [top](#symbols) | :leftwards_arrow_with_hook: | \`:leftwards_arrow_with_hook:\` | :arrow_right_hook: | \`:arrow_right_hook:\` | [top](#table-of-contents) |
+    | [top](#symbols) | :arrow_heading_up: | \`:arrow_heading_up:\` | :arrow_heading_down: | \`:arrow_heading_down:\` | [top](#table-of-contents) |
+    | [top](#symbols) | :arrows_clockwise: | \`:arrows_clockwise:\` | :arrows_counterclockwise: | \`:arrows_counterclockwise:\` | [top](#table-of-contents) |
+    | [top](#symbols) | :back: | \`:back:\` | :end: | \`:end:\` | [top](#table-of-contents) |
+    | [top](#symbols) | :on: | \`:on:\` | :soon: | \`:soon:\` | [top](#table-of-contents) |
+    | [top](#symbols) | :top: | \`:top:\` | | | [top](#table-of-contents) |
+
+    #### Religion
+
+    | | ico | shortcode | ico | shortcode | |
+    | - | :-: | - | :-: | - | - |
+    | [top](#symbols) | :place_of_worship: | \`:place_of_worship:\` | :atom_symbol: | \`:atom_symbol:\` | [top](#table-of-contents) |
+    | [top](#symbols) | :om: | \`:om:\` | :star_of_david: | \`:star_of_david:\` | [top](#table-of-contents) |
+    | [top](#symbols) | :wheel_of_dharma: | \`:wheel_of_dharma:\` | :yin_yang: | \`:yin_yang:\` | [top](#table-of-contents) |
+    | [top](#symbols) | :latin_cross: | \`:latin_cross:\` | :orthodox_cross: | \`:orthodox_cross:\` | [top](#table-of-contents) |
+    | [top](#symbols) | :star_and_crescent: | \`:star_and_crescent:\` | :peace_symbol: | \`:peace_symbol:\` | [top](#table-of-contents) |
+    | [top](#symbols) | :menorah: | \`:menorah:\` | :six_pointed_star: | \`:six_pointed_star:\` | [top](#table-of-contents) |
+
+    #### Zodiac
+
+    | | ico | shortcode | ico | shortcode | |
+    | - | :-: | - | :-: | - | - |
+    | [top](#symbols) | :aries: | \`:aries:\` | :taurus: | \`:taurus:\` | [top](#table-of-contents) |
+    | [top](#symbols) | :gemini: | \`:gemini:\` | :cancer: | \`:cancer:\` | [top](#table-of-contents) |
+    | [top](#symbols) | :leo: | \`:leo:\` | :virgo: | \`:virgo:\` | [top](#table-of-contents) |
+    | [top](#symbols) | :libra: | \`:libra:\` | :scorpius: | \`:scorpius:\` | [top](#table-of-contents) |
+    | [top](#symbols) | :sagittarius: | \`:sagittarius:\` | :capricorn: | \`:capricorn:\` | [top](#table-of-contents) |
+    | [top](#symbols) | :aquarius: | \`:aquarius:\` | :pisces: | \`:pisces:\` | [top](#table-of-contents) |
+    | [top](#symbols) | :ophiuchus: | \`:ophiuchus:\` | | | [top](#table-of-contents) |
+
+    #### Av Symbol
+
+    | | ico | shortcode | ico | shortcode | |
+    | - | :-: | - | :-: | - | - |
+    | [top](#symbols) | :twisted_rightwards_arrows: | \`:twisted_rightwards_arrows:\` | :repeat: | \`:repeat:\` | [top](#table-of-contents) |
+    | [top](#symbols) | :repeat_one: | \`:repeat_one:\` | :arrow_forward: | \`:arrow_forward:\` | [top](#table-of-contents) |
+    | [top](#symbols) | :fast_forward: | \`:fast_forward:\` | :next_track_button: | \`:next_track_button:\` | [top](#table-of-contents) |
+    | [top](#symbols) | :play_or_pause_button: | \`:play_or_pause_button:\` | :arrow_backward: | \`:arrow_backward:\` | [top](#table-of-contents) |
+    | [top](#symbols) | :rewind: | \`:rewind:\` | :previous_track_button: | \`:previous_track_button:\` | [top](#table-of-contents) |
+    | [top](#symbols) | :arrow_up_small: | \`:arrow_up_small:\` | :arrow_double_up: | \`:arrow_double_up:\` | [top](#table-of-contents) |
+    | [top](#symbols) | :arrow_down_small: | \`:arrow_down_small:\` | :arrow_double_down: | \`:arrow_double_down:\` | [top](#table-of-contents) |
+    | [top](#symbols) | :pause_button: | \`:pause_button:\` | :stop_button: | \`:stop_button:\` | [top](#table-of-contents) |
+    | [top](#symbols) | :record_button: | \`:record_button:\` | :cinema: | \`:cinema:\` | [top](#table-of-contents) |
+    | [top](#symbols) | :low_brightness: | \`:low_brightness:\` | :high_brightness: | \`:high_brightness:\` | [top](#table-of-contents) |
+    | [top](#symbols) | :signal_strength: | \`:signal_strength:\` | :vibration_mode: | \`:vibration_mode:\` | [top](#table-of-contents) |
+    | [top](#symbols) | :mobile_phone_off: | \`:mobile_phone_off:\` | | | [top](#table-of-contents) |
+
+    #### Other Symbol
+
+    | | ico | shortcode | ico | shortcode | |
+    | - | :-: | - | :-: | - | - |
+    | [top](#symbols) | :recycle: | \`:recycle:\` | :fleur_de_lis: | \`:fleur_de_lis:\` | [top](#table-of-contents) |
+    | [top](#symbols) | :trident: | \`:trident:\` | :name_badge: | \`:name_badge:\` | [top](#table-of-contents) |
+    | [top](#symbols) | :beginner: | \`:beginner:\` | :o: | \`:o:\` | [top](#table-of-contents) |
+    | [top](#symbols) | :white_check_mark: | \`:white_check_mark:\` | :ballot_box_with_check: | \`:ballot_box_with_check:\` | [top](#table-of-contents) |
+    | [top](#symbols) | :heavy_check_mark: | \`:heavy_check_mark:\` | :heavy_multiplication_x: | \`:heavy_multiplication_x:\` | [top](#table-of-contents) |
+    | [top](#symbols) | :x: | \`:x:\` | :negative_squared_cross_mark: | \`:negative_squared_cross_mark:\` | [top](#table-of-contents) |
+    | [top](#symbols) | :heavy_plus_sign: | \`:heavy_plus_sign:\` | :heavy_minus_sign: | \`:heavy_minus_sign:\` | [top](#table-of-contents) |
+    | [top](#symbols) | :heavy_division_sign: | \`:heavy_division_sign:\` | :curly_loop: | \`:curly_loop:\` | [top](#table-of-contents) |
+    | [top](#symbols) | :loop: | \`:loop:\` | :part_alternation_mark: | \`:part_alternation_mark:\` | [top](#table-of-contents) |
+    | [top](#symbols) | :eight_spoked_asterisk: | \`:eight_spoked_asterisk:\` | :eight_pointed_black_star: | \`:eight_pointed_black_star:\` | [top](#table-of-contents) |
+    | [top](#symbols) | :sparkle: | \`:sparkle:\` | :bangbang: | \`:bangbang:\` | [top](#table-of-contents) |
+    | [top](#symbols) | :interrobang: | \`:interrobang:\` | :question: | \`:question:\` | [top](#table-of-contents) |
+    | [top](#symbols) | :grey_question: | \`:grey_question:\` | :grey_exclamation: | \`:grey_exclamation:\` | [top](#table-of-contents) |
+    | [top](#symbols) | :exclamation: | \`:exclamation:\` <br /> \`:heavy_exclamation_mark:\` | :wavy_dash: | \`:wavy_dash:\` | [top](#table-of-contents) |
+    | [top](#symbols) | :copyright: | \`:copyright:\` | :registered: | \`:registered:\` | [top](#table-of-contents) |
+    | [top](#symbols) | :tm: | \`:tm:\` | | | [top](#table-of-contents) |
+
+    #### Keycap
+
+    | | ico | shortcode | ico | shortcode | |
+    | - | :-: | - | :-: | - | - |
+    | [top](#symbols) | :hash: | \`:hash:\` | :asterisk: | \`:asterisk:\` | [top](#table-of-contents) |
+    | [top](#symbols) | :zero: | \`:zero:\` | :one: | \`:one:\` | [top](#table-of-contents) |
+    | [top](#symbols) | :two: | \`:two:\` | :three: | \`:three:\` | [top](#table-of-contents) |
+    | [top](#symbols) | :four: | \`:four:\` | :five: | \`:five:\` | [top](#table-of-contents) |
+    | [top](#symbols) | :six: | \`:six:\` | :seven: | \`:seven:\` | [top](#table-of-contents) |
+    | [top](#symbols) | :eight: | \`:eight:\` | :nine: | \`:nine:\` | [top](#table-of-contents) |
+    | [top](#symbols) | :keycap_ten: | \`:keycap_ten:\` | | | [top](#table-of-contents) |
+
+    #### Alphanum
+
+    | | ico | shortcode | ico | shortcode | |
+    | - | :-: | - | :-: | - | - |
+    | [top](#symbols) | :capital_abcd: | \`:capital_abcd:\` | :abcd: | \`:abcd:\` | [top](#table-of-contents) |
+    | [top](#symbols) | :1234: | \`:1234:\` | :symbols: | \`:symbols:\` | [top](#table-of-contents) |
+    | [top](#symbols) | :abc: | \`:abc:\` | :a: | \`:a:\` | [top](#table-of-contents) |
+    | [top](#symbols) | :ab: | \`:ab:\` | :b: | \`:b:\` | [top](#table-of-contents) |
+    | [top](#symbols) | :cl: | \`:cl:\` | :cool: | \`:cool:\` | [top](#table-of-contents) |
+    | [top](#symbols) | :free: | \`:free:\` | :information_source: | \`:information_source:\` | [top](#table-of-contents) |
+    | [top](#symbols) | :id: | \`:id:\` | :m: | \`:m:\` | [top](#table-of-contents) |
+    | [top](#symbols) | :new: | \`:new:\` | :ng: | \`:ng:\` | [top](#table-of-contents) |
+    | [top](#symbols) | :o2: | \`:o2:\` | :ok: | \`:ok:\` | [top](#table-of-contents) |
+    | [top](#symbols) | :parking: | \`:parking:\` | :sos: | \`:sos:\` | [top](#table-of-contents) |
+    | [top](#symbols) | :up: | \`:up:\` | :vs: | \`:vs:\` | [top](#table-of-contents) |
+    | [top](#symbols) | :koko: | \`:koko:\` | :sa: | \`:sa:\` | [top](#table-of-contents) |
+    | [top](#symbols) | :u6708: | \`:u6708:\` | :u6709: | \`:u6709:\` | [top](#table-of-contents) |
+    | [top](#symbols) | :u6307: | \`:u6307:\` | :ideograph_advantage: | \`:ideograph_advantage:\` | [top](#table-of-contents) |
+    | [top](#symbols) | :u5272: | \`:u5272:\` | :u7121: | \`:u7121:\` | [top](#table-of-contents) |
+    | [top](#symbols) | :u7981: | \`:u7981:\` | :accept: | \`:accept:\` | [top](#table-of-contents) |
+    | [top](#symbols) | :u7533: | \`:u7533:\` | :u5408: | \`:u5408:\` | [top](#table-of-contents) |
+    | [top](#symbols) | :u7a7a: | \`:u7a7a:\` | :congratulations: | \`:congratulations:\` | [top](#table-of-contents) |
+    | [top](#symbols) | :secret: | \`:secret:\` | :u55b6: | \`:u55b6:\` | [top](#table-of-contents) |
+    | [top](#symbols) | :u6e80: | \`:u6e80:\` | | | [top](#table-of-contents) |
+
+    #### Geometric
+
+    | | ico | shortcode | ico | shortcode | |
+    | - | :-: | - | :-: | - | - |
+    | [top](#symbols) | :red_circle: | \`:red_circle:\` | :large_blue_circle: | \`:large_blue_circle:\` | [top](#table-of-contents) |
+    | [top](#symbols) | :black_circle: | \`:black_circle:\` | :white_circle: | \`:white_circle:\` | [top](#table-of-contents) |
+    | [top](#symbols) | :black_large_square: | \`:black_large_square:\` | :white_large_square: | \`:white_large_square:\` | [top](#table-of-contents) |
+    | [top](#symbols) | :black_medium_square: | \`:black_medium_square:\` | :white_medium_square: | \`:white_medium_square:\` | [top](#table-of-contents) |
+    | [top](#symbols) | :black_medium_small_square: | \`:black_medium_small_square:\` | :white_medium_small_square: | \`:white_medium_small_square:\` | [top](#table-of-contents) |
+    | [top](#symbols) | :black_small_square: | \`:black_small_square:\` | :white_small_square: | \`:white_small_square:\` | [top](#table-of-contents) |
+    | [top](#symbols) | :large_orange_diamond: | \`:large_orange_diamond:\` | :large_blue_diamond: | \`:large_blue_diamond:\` | [top](#table-of-contents) |
+    | [top](#symbols) | :small_orange_diamond: | \`:small_orange_diamond:\` | :small_blue_diamond: | \`:small_blue_diamond:\` | [top](#table-of-contents) |
+    | [top](#symbols) | :small_red_triangle: | \`:small_red_triangle:\` | :small_red_triangle_down: | \`:small_red_triangle_down:\` | [top](#table-of-contents) |
+    | [top](#symbols) | :diamond_shape_with_a_dot_inside: | \`:diamond_shape_with_a_dot_inside:\` | :radio_button: | \`:radio_button:\` | [top](#table-of-contents) |
+    | [top](#symbols) | :white_square_button: | \`:white_square_button:\` | :black_square_button: | \`:black_square_button:\` | [top](#table-of-contents) |
+
+    ### Flags
+
+    - [Flag](#flag)
+    - [Country Flag](#country-flag)
+
+    #### Flag
+
+    | | ico | shortcode | ico | shortcode | |
+    | - | :-: | - | :-: | - | - |
+    | [top](#flags) | :checkered_flag: | \`:checkered_flag:\` | :triangular_flag_on_post: | \`:triangular_flag_on_post:\` | [top](#table-of-contents) |
+    | [top](#flags) | :crossed_flags: | \`:crossed_flags:\` | :black_flag: | \`:black_flag:\` | [top](#table-of-contents) |
+    | [top](#flags) | :white_flag: | \`:white_flag:\` | :rainbow_flag: | \`:rainbow_flag:\` | [top](#table-of-contents) |
+
+    #### Country Flag
+
+    | | ico | shortcode | ico | shortcode | |
+    | - | :-: | - | :-: | - | - |
+    | [top](#flags) | :andorra: | \`:andorra:\` | :united_arab_emirates: | \`:united_arab_emirates:\` | [top](#table-of-contents) |
+    | [top](#flags) | :afghanistan: | \`:afghanistan:\` | :antigua_barbuda: | \`:antigua_barbuda:\` | [top](#table-of-contents) |
+    | [top](#flags) | :anguilla: | \`:anguilla:\` | :albania: | \`:albania:\` | [top](#table-of-contents) |
+    | [top](#flags) | :armenia: | \`:armenia:\` | :angola: | \`:angola:\` | [top](#table-of-contents) |
+    | [top](#flags) | :antarctica: | \`:antarctica:\` | :argentina: | \`:argentina:\` | [top](#table-of-contents) |
+    | [top](#flags) | :american_samoa: | \`:american_samoa:\` | :austria: | \`:austria:\` | [top](#table-of-contents) |
+    | [top](#flags) | :australia: | \`:australia:\` | :aruba: | \`:aruba:\` | [top](#table-of-contents) |
+    | [top](#flags) | :aland_islands: | \`:aland_islands:\` | :azerbaijan: | \`:azerbaijan:\` | [top](#table-of-contents) |
+    | [top](#flags) | :bosnia_herzegovina: | \`:bosnia_herzegovina:\` | :barbados: | \`:barbados:\` | [top](#table-of-contents) |
+    | [top](#flags) | :bangladesh: | \`:bangladesh:\` | :belgium: | \`:belgium:\` | [top](#table-of-contents) |
+    | [top](#flags) | :burkina_faso: | \`:burkina_faso:\` | :bulgaria: | \`:bulgaria:\` | [top](#table-of-contents) |
+    | [top](#flags) | :bahrain: | \`:bahrain:\` | :burundi: | \`:burundi:\` | [top](#table-of-contents) |
+    | [top](#flags) | :benin: | \`:benin:\` | :st_barthelemy: | \`:st_barthelemy:\` | [top](#table-of-contents) |
+    | [top](#flags) | :bermuda: | \`:bermuda:\` | :brunei: | \`:brunei:\` | [top](#table-of-contents) |
+    | [top](#flags) | :bolivia: | \`:bolivia:\` | :caribbean_netherlands: | \`:caribbean_netherlands:\` | [top](#table-of-contents) |
+    | [top](#flags) | :brazil: | \`:brazil:\` | :bahamas: | \`:bahamas:\` | [top](#table-of-contents) |
+    | [top](#flags) | :bhutan: | \`:bhutan:\` | :botswana: | \`:botswana:\` | [top](#table-of-contents) |
+    | [top](#flags) | :belarus: | \`:belarus:\` | :belize: | \`:belize:\` | [top](#table-of-contents) |
+    | [top](#flags) | :canada: | \`:canada:\` | :cocos_islands: | \`:cocos_islands:\` | [top](#table-of-contents) |
+    | [top](#flags) | :congo_kinshasa: | \`:congo_kinshasa:\` | :central_african_republic: | \`:central_african_republic:\` | [top](#table-of-contents) |
+    | [top](#flags) | :congo_brazzaville: | \`:congo_brazzaville:\` | :switzerland: | \`:switzerland:\` | [top](#table-of-contents) |
+    | [top](#flags) | :cote_divoire: | \`:cote_divoire:\` | :cook_islands: | \`:cook_islands:\` | [top](#table-of-contents) |
+    | [top](#flags) | :chile: | \`:chile:\` | :cameroon: | \`:cameroon:\` | [top](#table-of-contents) |
+    | [top](#flags) | :cn: | \`:cn:\` | :colombia: | \`:colombia:\` | [top](#table-of-contents) |
+    | [top](#flags) | :costa_rica: | \`:costa_rica:\` | :cuba: | \`:cuba:\` | [top](#table-of-contents) |
+    | [top](#flags) | :cape_verde: | \`:cape_verde:\` | :curacao: | \`:curacao:\` | [top](#table-of-contents) |
+    | [top](#flags) | :christmas_island: | \`:christmas_island:\` | :cyprus: | \`:cyprus:\` | [top](#table-of-contents) |
+    | [top](#flags) | :czech_republic: | \`:czech_republic:\` | :de: | \`:de:\` | [top](#table-of-contents) |
+    | [top](#flags) | :djibouti: | \`:djibouti:\` | :denmark: | \`:denmark:\` | [top](#table-of-contents) |
+    | [top](#flags) | :dominica: | \`:dominica:\` | :dominican_republic: | \`:dominican_republic:\` | [top](#table-of-contents) |
+    | [top](#flags) | :algeria: | \`:algeria:\` | :ecuador: | \`:ecuador:\` | [top](#table-of-contents) |
+    | [top](#flags) | :estonia: | \`:estonia:\` | :egypt: | \`:egypt:\` | [top](#table-of-contents) |
+    | [top](#flags) | :western_sahara: | \`:western_sahara:\` | :eritrea: | \`:eritrea:\` | [top](#table-of-contents) |
+    | [top](#flags) | :es: | \`:es:\` | :ethiopia: | \`:ethiopia:\` | [top](#table-of-contents) |
+    | [top](#flags) | :eu: | \`:eu:\` <br /> \`:european_union:\` | :finland: | \`:finland:\` | [top](#table-of-contents) |
+    | [top](#flags) | :fiji: | \`:fiji:\` | :falkland_islands: | \`:falkland_islands:\` | [top](#table-of-contents) |
+    | [top](#flags) | :micronesia: | \`:micronesia:\` | :faroe_islands: | \`:faroe_islands:\` | [top](#table-of-contents) |
+    | [top](#flags) | :fr: | \`:fr:\` | :gabon: | \`:gabon:\` | [top](#table-of-contents) |
+    | [top](#flags) | :gb: | \`:gb:\` <br /> \`:uk:\` | :grenada: | \`:grenada:\` | [top](#table-of-contents) |
+    | [top](#flags) | :georgia: | \`:georgia:\` | :french_guiana: | \`:french_guiana:\` | [top](#table-of-contents) |
+    | [top](#flags) | :guernsey: | \`:guernsey:\` | :ghana: | \`:ghana:\` | [top](#table-of-contents) |
+    | [top](#flags) | :gibraltar: | \`:gibraltar:\` | :greenland: | \`:greenland:\` | [top](#table-of-contents) |
+    | [top](#flags) | :gambia: | \`:gambia:\` | :guinea: | \`:guinea:\` | [top](#table-of-contents) |
+    | [top](#flags) | :guadeloupe: | \`:guadeloupe:\` | :equatorial_guinea: | \`:equatorial_guinea:\` | [top](#table-of-contents) |
+    | [top](#flags) | :greece: | \`:greece:\` | :south_georgia_south_sandwich_islands: | \`:south_georgia_south_sandwich_islands:\` | [top](#table-of-contents) |
+    | [top](#flags) | :guatemala: | \`:guatemala:\` | :guam: | \`:guam:\` | [top](#table-of-contents) |
+    | [top](#flags) | :guinea_bissau: | \`:guinea_bissau:\` | :guyana: | \`:guyana:\` | [top](#table-of-contents) |
+    | [top](#flags) | :hong_kong: | \`:hong_kong:\` | :honduras: | \`:honduras:\` | [top](#table-of-contents) |
+    | [top](#flags) | :croatia: | \`:croatia:\` | :haiti: | \`:haiti:\` | [top](#table-of-contents) |
+    | [top](#flags) | :hungary: | \`:hungary:\` | :canary_islands: | \`:canary_islands:\` | [top](#table-of-contents) |
+    | [top](#flags) | :indonesia: | \`:indonesia:\` | :ireland: | \`:ireland:\` | [top](#table-of-contents) |
+    | [top](#flags) | :israel: | \`:israel:\` | :isle_of_man: | \`:isle_of_man:\` | [top](#table-of-contents) |
+    | [top](#flags) | :india: | \`:india:\` | :british_indian_ocean_territory: | \`:british_indian_ocean_territory:\` | [top](#table-of-contents) |
+    | [top](#flags) | :iraq: | \`:iraq:\` | :iran: | \`:iran:\` | [top](#table-of-contents) |
+    | [top](#flags) | :iceland: | \`:iceland:\` | :it: | \`:it:\` | [top](#table-of-contents) |
+    | [top](#flags) | :jersey: | \`:jersey:\` | :jamaica: | \`:jamaica:\` | [top](#table-of-contents) |
+    | [top](#flags) | :jordan: | \`:jordan:\` | :jp: | \`:jp:\` | [top](#table-of-contents) |
+    | [top](#flags) | :kenya: | \`:kenya:\` | :kyrgyzstan: | \`:kyrgyzstan:\` | [top](#table-of-contents) |
+    | [top](#flags) | :cambodia: | \`:cambodia:\` | :kiribati: | \`:kiribati:\` | [top](#table-of-contents) |
+    | [top](#flags) | :comoros: | \`:comoros:\` | :st_kitts_nevis: | \`:st_kitts_nevis:\` | [top](#table-of-contents) |
+    | [top](#flags) | :north_korea: | \`:north_korea:\` | :kr: | \`:kr:\` | [top](#table-of-contents) |
+    | [top](#flags) | :kuwait: | \`:kuwait:\` | :cayman_islands: | \`:cayman_islands:\` | [top](#table-of-contents) |
+    | [top](#flags) | :kazakhstan: | \`:kazakhstan:\` | :laos: | \`:laos:\` | [top](#table-of-contents) |
+    | [top](#flags) | :lebanon: | \`:lebanon:\` | :st_lucia: | \`:st_lucia:\` | [top](#table-of-contents) |
+    | [top](#flags) | :liechtenstein: | \`:liechtenstein:\` | :sri_lanka: | \`:sri_lanka:\` | [top](#table-of-contents) |
+    | [top](#flags) | :liberia: | \`:liberia:\` | :lesotho: | \`:lesotho:\` | [top](#table-of-contents) |
+    | [top](#flags) | :lithuania: | \`:lithuania:\` | :luxembourg: | \`:luxembourg:\` | [top](#table-of-contents) |
+    | [top](#flags) | :latvia: | \`:latvia:\` | :libya: | \`:libya:\` | [top](#table-of-contents) |
+    | [top](#flags) | :morocco: | \`:morocco:\` | :monaco: | \`:monaco:\` | [top](#table-of-contents) |
+    | [top](#flags) | :moldova: | \`:moldova:\` | :montenegro: | \`:montenegro:\` | [top](#table-of-contents) |
+    | [top](#flags) | :madagascar: | \`:madagascar:\` | :marshall_islands: | \`:marshall_islands:\` | [top](#table-of-contents) |
+    | [top](#flags) | :macedonia: | \`:macedonia:\` | :mali: | \`:mali:\` | [top](#table-of-contents) |
+    | [top](#flags) | :myanmar: | \`:myanmar:\` | :mongolia: | \`:mongolia:\` | [top](#table-of-contents) |
+    | [top](#flags) | :macau: | \`:macau:\` | :northern_mariana_islands: | \`:northern_mariana_islands:\` | [top](#table-of-contents) |
+    | [top](#flags) | :martinique: | \`:martinique:\` | :mauritania: | \`:mauritania:\` | [top](#table-of-contents) |
+    | [top](#flags) | :montserrat: | \`:montserrat:\` | :malta: | \`:malta:\` | [top](#table-of-contents) |
+    | [top](#flags) | :mauritius: | \`:mauritius:\` | :maldives: | \`:maldives:\` | [top](#table-of-contents) |
+    | [top](#flags) | :malawi: | \`:malawi:\` | :mexico: | \`:mexico:\` | [top](#table-of-contents) |
+    | [top](#flags) | :malaysia: | \`:malaysia:\` | :mozambique: | \`:mozambique:\` | [top](#table-of-contents) |
+    | [top](#flags) | :namibia: | \`:namibia:\` | :new_caledonia: | \`:new_caledonia:\` | [top](#table-of-contents) |
+    | [top](#flags) | :niger: | \`:niger:\` | :norfolk_island: | \`:norfolk_island:\` | [top](#table-of-contents) |
+    | [top](#flags) | :nigeria: | \`:nigeria:\` | :nicaragua: | \`:nicaragua:\` | [top](#table-of-contents) |
+    | [top](#flags) | :netherlands: | \`:netherlands:\` | :norway: | \`:norway:\` | [top](#table-of-contents) |
+    | [top](#flags) | :nepal: | \`:nepal:\` | :nauru: | \`:nauru:\` | [top](#table-of-contents) |
+    | [top](#flags) | :niue: | \`:niue:\` | :new_zealand: | \`:new_zealand:\` | [top](#table-of-contents) |
+    | [top](#flags) | :oman: | \`:oman:\` | :panama: | \`:panama:\` | [top](#table-of-contents) |
+    | [top](#flags) | :peru: | \`:peru:\` | :french_polynesia: | \`:french_polynesia:\` | [top](#table-of-contents) |
+    | [top](#flags) | :papua_new_guinea: | \`:papua_new_guinea:\` | :philippines: | \`:philippines:\` | [top](#table-of-contents) |
+    | [top](#flags) | :pakistan: | \`:pakistan:\` | :poland: | \`:poland:\` | [top](#table-of-contents) |
+    | [top](#flags) | :st_pierre_miquelon: | \`:st_pierre_miquelon:\` | :pitcairn_islands: | \`:pitcairn_islands:\` | [top](#table-of-contents) |
+    | [top](#flags) | :puerto_rico: | \`:puerto_rico:\` | :palestinian_territories: | \`:palestinian_territories:\` | [top](#table-of-contents) |
+    | [top](#flags) | :portugal: | \`:portugal:\` | :palau: | \`:palau:\` | [top](#table-of-contents) |
+    | [top](#flags) | :paraguay: | \`:paraguay:\` | :qatar: | \`:qatar:\` | [top](#table-of-contents) |
+    | [top](#flags) | :reunion: | \`:reunion:\` | :romania: | \`:romania:\` | [top](#table-of-contents) |
+    | [top](#flags) | :serbia: | \`:serbia:\` | :ru: | \`:ru:\` | [top](#table-of-contents) |
+    | [top](#flags) | :rwanda: | \`:rwanda:\` | :saudi_arabia: | \`:saudi_arabia:\` | [top](#table-of-contents) |
+    | [top](#flags) | :solomon_islands: | \`:solomon_islands:\` | :seychelles: | \`:seychelles:\` | [top](#table-of-contents) |
+    | [top](#flags) | :sudan: | \`:sudan:\` | :sweden: | \`:sweden:\` | [top](#table-of-contents) |
+    | [top](#flags) | :singapore: | \`:singapore:\` | :st_helena: | \`:st_helena:\` | [top](#table-of-contents) |
+    | [top](#flags) | :slovenia: | \`:slovenia:\` | :slovakia: | \`:slovakia:\` | [top](#table-of-contents) |
+    | [top](#flags) | :sierra_leone: | \`:sierra_leone:\` | :san_marino: | \`:san_marino:\` | [top](#table-of-contents) |
+    | [top](#flags) | :senegal: | \`:senegal:\` | :somalia: | \`:somalia:\` | [top](#table-of-contents) |
+    | [top](#flags) | :suriname: | \`:suriname:\` | :south_sudan: | \`:south_sudan:\` | [top](#table-of-contents) |
+    | [top](#flags) | :sao_tome_principe: | \`:sao_tome_principe:\` | :el_salvador: | \`:el_salvador:\` | [top](#table-of-contents) |
+    | [top](#flags) | :sint_maarten: | \`:sint_maarten:\` | :syria: | \`:syria:\` | [top](#table-of-contents) |
+    | [top](#flags) | :swaziland: | \`:swaziland:\` | :turks_caicos_islands: | \`:turks_caicos_islands:\` | [top](#table-of-contents) |
+    | [top](#flags) | :chad: | \`:chad:\` | :french_southern_territories: | \`:french_southern_territories:\` | [top](#table-of-contents) |
+    | [top](#flags) | :togo: | \`:togo:\` | :thailand: | \`:thailand:\` | [top](#table-of-contents) |
+    | [top](#flags) | :tajikistan: | \`:tajikistan:\` | :tokelau: | \`:tokelau:\` | [top](#table-of-contents) |
+    | [top](#flags) | :timor_leste: | \`:timor_leste:\` | :turkmenistan: | \`:turkmenistan:\` | [top](#table-of-contents) |
+    | [top](#flags) | :tunisia: | \`:tunisia:\` | :tonga: | \`:tonga:\` | [top](#table-of-contents) |
+    | [top](#flags) | :tr: | \`:tr:\` | :trinidad_tobago: | \`:trinidad_tobago:\` | [top](#table-of-contents) |
+    | [top](#flags) | :tuvalu: | \`:tuvalu:\` | :taiwan: | \`:taiwan:\` | [top](#table-of-contents) |
+    | [top](#flags) | :tanzania: | \`:tanzania:\` | :ukraine: | \`:ukraine:\` | [top](#table-of-contents) |
+    | [top](#flags) | :uganda: | \`:uganda:\` | :us: | \`:us:\` | [top](#table-of-contents) |
+    | [top](#flags) | :uruguay: | \`:uruguay:\` | :uzbekistan: | \`:uzbekistan:\` | [top](#table-of-contents) |
+    | [top](#flags) | :vatican_city: | \`:vatican_city:\` | :st_vincent_grenadines: | \`:st_vincent_grenadines:\` | [top](#table-of-contents) |
+    | [top](#flags) | :venezuela: | \`:venezuela:\` | :british_virgin_islands: | \`:british_virgin_islands:\` | [top](#table-of-contents) |
+    | [top](#flags) | :us_virgin_islands: | \`:us_virgin_islands:\` | :vietnam: | \`:vietnam:\` | [top](#table-of-contents) |
+    | [top](#flags) | :vanuatu: | \`:vanuatu:\` | :wallis_futuna: | \`:wallis_futuna:\` | [top](#table-of-contents) |
+    | [top](#flags) | :samoa: | \`:samoa:\` | :kosovo: | \`:kosovo:\` | [top](#table-of-contents) |
+    | [top](#flags) | :yemen: | \`:yemen:\` | :mayotte: | \`:mayotte:\` | [top](#table-of-contents) |
+    | [top](#flags) | :south_africa: | \`:south_africa:\` | :zambia: | \`:zambia:\` | [top](#table-of-contents) |
+    | [top](#flags) | :zimbabwe: | \`:zimbabwe:\` | | | [top](#table-of-contents) |
+
+    ### GitHub Custom Emoji
+
+    | | ico | shortcode | ico | shortcode | |
+    | - | :-: | - | :-: | - | - |
+    | [top](#github-custom-emoji) | :atom: | \`:atom:\` | :basecamp: | \`:basecamp:\` | [top](#table-of-contents) |
+    | [top](#github-custom-emoji) | :basecampy: | \`:basecampy:\` | :bowtie: | \`:bowtie:\` | [top](#table-of-contents) |
+    | [top](#github-custom-emoji) | :electron: | \`:electron:\` | :feelsgood: | \`:feelsgood:\` | [top](#table-of-contents) |
+    | [top](#github-custom-emoji) | :finnadie: | \`:finnadie:\` | :goberserk: | \`:goberserk:\` | [top](#table-of-contents) |
+    | [top](#github-custom-emoji) | :godmode: | \`:godmode:\` | :hurtrealbad: | \`:hurtrealbad:\` | [top](#table-of-contents) |
+    | [top](#github-custom-emoji) | :neckbeard: | \`:neckbeard:\` | :octocat: | \`:octocat:\` | [top](#table-of-contents) |
+    | [top](#github-custom-emoji) | :rage1: | \`:rage1:\` | :rage2: | \`:rage2:\` | [top](#table-of-contents) |
+    | [top](#github-custom-emoji) | :rage3: | \`:rage3:\` | :rage4: | \`:rage4:\` | [top](#table-of-contents) |
+    | [top](#github-custom-emoji) | :shipit: | \`:shipit:\` <br /> \`:squirrel:\` | :suspect: | \`:suspect:\` | [top](#table-of-contents) |
+    | [top](#github-custom-emoji) | :trollface: | \`:trollface:\` | | | [top](#table-of-contents) |
+    "
+  `);
+});
